% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{addCovariatesToMetricsDf}
\alias{addCovariatesToMetricsDf}
\title{Add donor level covariates to the single cell metrics data frame}
\usage{
addCovariatesToMetricsDf(metricsDF, donorCovariatesFile)
}
\arguments{
\item{metricsDF}{A data frame containing cell level metrics.  Must contain a DONOR column.}

\item{donorCovariatesFile}{A donor covariates file in the eQTL format}
}
\value{
The metrics file with the additonal donor level metrics
}
\description{
Add donor level covariates to the single cell metrics data frame
}
