% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{addGeneLocations}
\alias{addGeneLocations}
\title{Add the gene location of the gene to the data frame.
Look up the gene name in the to-be-annotatated data frame using the geneColName}
\usage{
addGeneLocations(reducedGTF = NULL, df, geneColName = "gene")
}
\arguments{
\item{reducedGTF}{A data.frame of gene locations.  Each row is a gene, columns are geneid (gene symbol), chr (contig), s1 (start), s2 (end).
If this parameter is null the operation is a no-op and the original dataframe is returned.}

\item{df}{The data table to annotate}

\item{geneColName}{The column name in df that contains the gene name.}
}
\description{
Add the gene location of the gene to the data frame.
Look up the gene name in the to-be-annotatated data frame using the geneColName
}
