% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneSNPPositions.R
\name{addLocationAnnotationToGenes}
\alias{addLocationAnnotationToGenes}
\title{Adds the gene location to an eQTL result data frame.}
\usage{
addLocationAnnotationToGenes(eQTLData, geneLoc)
}
\arguments{
\item{eQTLData}{a data frame that has a column "gene" containing the gene ID.}

\item{geneLoc}{Data frame of gene locations in the standard eQTL format containing the columns geneid,chr,s1,s2 that encode the snp name, chromosome, and start position, end position.}
}
\value{
The data frame with an additonal columns gene_start_pos and gene_end_pos
}
\description{
Adds the gene location to an eQTL result data frame.
}
