% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneSNPPositions.R
\name{addLocationAnnotationToSNPs}
\alias{addLocationAnnotationToSNPs}
\title{Adds the SNP location to an eQTL result data frame.}
\usage{
addLocationAnnotationToSNPs(eQTLData, snpLoc)
}
\arguments{
\item{eQTLData}{a data frame that has a column "SNP" containing the SNP ID.}

\item{snpLoc}{Data frame of SNP locations in the standard eQTL format containing the columns snp,chr,pos that encode the snp name, chromosome, and position.}
}
\value{
The data frame with an additional column snp_pos
}
\description{
Adds the SNP location to an eQTL result data frame.
}
