% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{addMedianExpression}
\alias{addMedianExpression}
\title{Add median expression per gene and effect size normalized to expression}
\usage{
addMedianExpression(result, gene)
}
\arguments{
\item{result}{The eQTL results}

\item{gene}{The gene expression data}
}
\value{
The eQTL results with added columns.
}
\description{
The effect size is the beta of the linear regression / median expression for that gene.
}
