% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{addMedianExpressionThis}
\alias{addMedianExpressionThis}
\title{Add the median expression of the gene to the data frame.
Look up the gene name in the to-be-annotatated data frame using the geneColName}
\usage{
addMedianExpressionThis(expressionData, df, geneColName = "gene")
}
\arguments{
\item{expressionData}{A matrix of expression data.  Each row is a gene, each column a donor.
Each rowname is a gene names.}

\item{df}{The data frame to annotate}

\item{geneColName}{The column name in df that contains the gene name.}
}
\description{
Add the median expression of the gene to the data frame.
Look up the gene name in the to-be-annotatated data frame using the geneColName
}
