% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchEqtlResults.R
\name{aggregateEqtlResults}
\alias{aggregateEqtlResults}
\title{Aggregate results across many eQTL data sets for a list SNPs/Genes pairs}
\usage{
aggregateEqtlResults(
  rootDir,
  queryFile,
  outFile,
  outSummaryFile,
  ldTargetPopulation = NULL
)
}
\arguments{
\item{rootDir}{A root directory under which to find eQTL files}

\item{queryFile}{A tab delimited file with 2 columns.  SNP: the rsID of the SNP of interest.
GENE: the gene of interest.  Each row corresponds to a SNP/Gene pair.  SNPs and genes may be
repeated multple times on different rows.}

\item{outFile}{The location to write the output file.  This file is fairly verbose,
with the explicit SNP/Gene pair and independent SNP and gene results on a single line.}

\item{outSummaryFile}{The location to write the summary file to.  This file contains the best (most significant
q-value or emperic p-value) for each SNP gene pair across all experiments, as well as the best
result for each SNP and gene independently.}

\item{ldTargetPopulation}{if supplied, use LDLinkR to add LD scores to the summary file for the "best gene" result.}
}
\description{
Given a root directory, recursively find all eQTL_result.txt files under that directory
For each, extract the empiric and permuted results for the SNP/Gene pairs of interest
as well as the best results for each SNP or gene independently
}
