% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_Example.R
\name{annotateEQTL}
\alias{annotateEQTL}
\title{eQTL discovery.}
\usage{
annotateEQTL(
  output_file_name_cis,
  cis_eqtl_file_name,
  params_file_name,
  SNP_file_name,
  snps_location_file_name,
  expression_file_name,
  gene_location_file_name,
  useQuantileNormalization = F,
  minorAlleleFreqThreshold = NULL,
  snp_gene_map_file_name = NULL
)
}
\arguments{
\item{output_file_name_cis}{Result written here.}

\item{cis_eqtl_file_name}{output produced by Matrix_eQTL_main(output_file_name.cis)}

\item{params_file_name}{Most Matrix_eQTL_main returned params, in tabular format.}

\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{snp_gene_map_file_name}{optional file with columns 'snp' and 'gene' that represents a many-to-many
map.  Note that currently this is only used to indicate that SNP names
have gene suffix that needs to be removed}
}
\description{
Annotates eQTLs produced by Matrix eQTL package.
}
\seealso{
\code{\link{Matrix_eQTL_main}}.

MatrixEQTL::Matrix_eQTL_main
}
