% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaApproximatatedPvalueCorrection.R
\name{betaApproximationPvalue}
\alias{betaApproximationPvalue}
\title{Compute a pvalue correction using a Beta approximation (fastQTL)}
\usage{
betaApproximationPvalue(empiricP, permutedPDist)
}
\arguments{
\item{empiricP}{The empiric pvalue to correct}

\item{permutedPDist}{A distribution of permuted pvalues.  This is used to estimate the shape of the beta distribution.
1000-10000 permutations should be sufficient for a well calibrated result.}
}
\value{
A data frame containing the direct permutation result and the beta approximation
}
\description{
Fit permuted p-value distribution to beta distribution.
This characterizes the extreme tail of the null without directly sampling it.
See the paper: https://academic.oup.com/bioinformatics/article/32/10/1479/1742545
See the code: https://github.com/broadinstitute/tensorqtl/blob/master/tensorqtl/core.py
}
