% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{binSignTestByMAF}
\alias{binSignTestByMAF}
\title{Generate sign test results for an eQTL data set as compared to a reference eQTL data set, binned by index SNP MAF.}
\usage{
binSignTestByMAF(
  permutedQTLFile,
  unfilteredQTLFile,
  qvalueThreshold = 0.05,
  binSize = 0.025
)
}
\arguments{
\item{permutedQTLFile}{An eQTL result file that has gone through permutation (or other gene-level correction) and has a q-value.}

\item{unfilteredQTLFile}{An eQTL result file run on all SNPs on the data set.  This data set should contain all SNP/Gene pairs in the permutedQTLFile,
even if they wouldn't normally be included in the data set.}

\item{qvalueThreshold}{A minimum q-value threshold for a gene from the permutedQTLFile to be included in the result set.}

\item{binSize}{The size of each bin of minor allele frequency.  This builds a sequence of bins from 0 to 0.5 by this size.
The sign test result is estimated for index SNPs within each bin.}
}
\value{
A data frame containing the MAF bins and sign test results at each tested bin.
}
\description{
Bin sign test results by minor allele frequency and generate useful plots.
}
\seealso{
DropSeq.eqtl::testEffectsSignsMatch
}
