% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{buildGWASeQTLSNPSets}
\alias{buildGWASeQTLSNPSets}
\title{For each eQTL that has a q-value <= qValueThreshold, get the gwas results and eQTL results and build a dataframe of
SNP results for the intersecting SNP set.
Calculates Z-scores for each SNP in both data sets.}
\usage{
buildGWASeQTLSNPSets(
  eQTLPermuatedFile,
  eQTLLargeWindowFile,
  gwasFile,
  qValueThreshold = 0.05,
  gwasPValueThreshold = 1e-06,
  geneListFile = NULL,
  mode = c("intersect", "union"),
  outFileDir,
  outManifestFile = NULL,
  verbose = F
)
}
\arguments{
\item{eQTLPermuatedFile}{the permuted eQTL results containing the q-value.}

\item{eQTLLargeWindowFile}{the empiric eQTL results at a large window size (250kb around gene)}

\item{gwasFile}{The GWAS results file, containing the following columns:
chr, snpid, a1, a2, pos, p.  These are chromosome, snp name, allele 1, allele 2, and p-value.}

\item{qValueThreshold}{only look at genes with qvalues <= this threshold.}

\item{gwasPValueThreshold}{only look at blocks that have a GWAS SNP with a value <= this threshold.}

\item{geneListFile}{override the qValueThreshold and output snp blocks for all genes in the list.}

\item{mode}{If union, builds a data frame containing all eQTL and GWAS SNPs, with the common snps "in frame".  If intersect, only use the SNPs present in both data sets.}

\item{outFileDir}{The directory to deposit files.}

\item{outManifestFile}{The location to write the manifest file, which contains the gene name and file name (relative to the outFileDir) containing the SNP dataframe.
If set to null, outputs the file name "Gene_Manifest.txt" to the outFileDir}

\item{verbose}{Write progress report to standard output.}
}
\description{
For each eQTL that has a q-value <= qValueThreshold, get the gwas results and eQTL results and build a dataframe of
SNP results for the intersecting SNP set.
Calculates Z-scores for each SNP in both data sets.
}
