% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{buildLDBlock}
\alias{buildLDBlock}
\title{Query a VCF to extract a square matrix of geneotype correlation.}
\usage{
buildLDBlock(
  df,
  vcfFile,
  outputBaseFileName,
  bcfToolsLocation,
  ldScoreBcftoolsPluginLocation,
  cleanUpFiles = F,
  dryRun,
  verbose = T
)
}
\arguments{
\item{df}{A data frame containing the GWAS/eQTL combined data.}

\item{vcfFile}{A VCF file to extract LD data from for this region.  Suggest using a population from thousand genomes or similar.}

\item{outputBaseFileName}{The prefix for all files for this region, including the full directory path.}

\item{bcfToolsLocation}{The full path to the bcftools executable.}

\item{ldScoreBcftoolsPluginLocation}{The ful path to the ldScore plugin for bcftools.}

\item{cleanUpFiles}{delete the intermediate files generated if true (saves space!)}

\item{dryRun}{If true, do not run bcftools call, only echo command string if verbose=T.}

\item{verbose}{Print the command string.}
}
\value{
A square matrix of genotype correlation.
}
\description{
Query a VCF to extract a square matrix of geneotype correlation.
}
