% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateGeneCorrelation.R
\name{buildScatterPlot}
\alias{buildScatterPlot}
\title{Builds a scatter plot}
\usage{
buildScatterPlot(
  permutedResultsFile,
  contigGroupsFile,
  colNames = c("SEX_permuted_p", "FRACTION_X_permuted_p"),
  dataLabels = c("sex permuted pval [-log10]", "fraction x permuted pval [-log10]"),
  useLog10 = T,
  useNegative = T,
  strTitle = ""
)
}
\arguments{
\item{permutedResultsFile}{permuted eQTL results file}

\item{contigGroupsFile}{A YAML file containing each contig name and groups it belongs to.  In this case the contigs should belong to the groups autosome and chrX.  If provided, renormalizes autosome expression data to not include the x chromosomne.}

\item{colNames}{Column names to plot}

\item{dataLabels}{Plot labels for column names}

\item{useLog10}{Set to true to log10 transform the data}

\item{useNegative}{Set true to multiply the results by -1}

\item{strTitle}{The title for plots.}
}
\description{
Builds a scatter plot
}
