% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePermutationResults.R
\name{combinePermutationResults}
\alias{combinePermutationResults}
\title{Combine permutation results from runAdaptivePermutationEQTL.}
\usage{
combinePermutationResults(inFileList, outFile, sequenceDictionaryFile = NULL)
}
\arguments{
\item{inFileList}{A list of eQTL permutation result files to combine.}

\item{outFile}{The single file to emit with permutation results across all chromosomes and FDR correction.}

\item{sequenceDictionaryFile}{A file containing the ordering of chromosomes in the reference.  If not set, do not sort output file.}
}
\description{
Merges results across the per-chromosome files into a single output file, then sorts the data in genomic order
and adds a q-value calculation to be select the final set of results.
}
