% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePermutationResults.R
\name{combinePermutationResultsFromBaseNames}
\alias{combinePermutationResultsFromBaseNames}
\title{Combine permutation results.}
\usage{
combinePermutationResultsFromBaseNames(
  inDir,
  pattern = "eQTL_results.full_permuted",
  outFile
)
}
\arguments{
\item{inDir}{The directory in which the per-chromosome permutation results are located}

\item{pattern}{The string prefix or suffix to filter file names by to get just the permutation result files.}

\item{outFile}{The single file to emit with permutation results across all chromosomes and FDR correction.}
}
\description{
Combine permutation results.
}
