% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{convertToSlicedData}
\alias{convertToSlicedData}
\title{Convert a data table into matrix eQTL's slice data format.}
\usage{
convertToSlicedData(dt, numRowsPerSlice = 1000)
}
\arguments{
\item{dt}{a data table in the expression data or gentotype data format.}

\item{numRowsPerSlice}{the number of rows per slice.}
}
\value{
A SlicedData object containing the data
}
\description{
Convert a data table into matrix eQTL's slice data format.
}
