% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{doAcrossResults}
\alias{doAcrossResults}
\title{Run a function across eQTL results}
\usage{
doAcrossResults(rootDirs, queryFile, FUN, usePeer)
}
\arguments{
\item{rootDirs}{character vector of paths to root directories containing eQTL results}

\item{queryFile}{A tab delimited file with 2 columns.  SNP: the rsID of the SNP of interest.}

\item{FUN}{function called across query pairs for each rootDirs[i] i=1:length(rootDirs) 
with signature function(gene = character(), geneExpression=numeric(), snp=character(), rsid=character(), 
genotype=integer(), files=fetchEqtlFiles(rootDirs[i]))}

\item{usePeer}{boolean flag to indicate whether or not to use peer-adjusted expression}
}
\description{
Run a function across eQTL results
}
