% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampleMetaCells.R
\name{downsampleMetaCells}
\alias{downsampleMetaCells}
\title{Downsample the UMIs of donors by some fraction of their UMIs.}
\usage{
downsampleMetaCells(metaCellFile, fractionUMIs, outFile = NULL)
}
\arguments{
\item{metaCellFile}{A meta-cell file}

\item{fractionUMIs}{the fraction of UMIs to retain}

\item{outFile}{The output file to write results to.  If null return the data frame.}
}
\value{
The meta cell matrix after downsampling.
}
\description{
Downsample the UMIs of donors by some fraction of their UMIs.
}
