% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{effectSizePlotSimple}
\alias{effectSizePlotSimple}
\title{A plot of effect sizes}
\usage{
effectSizePlotSimple(g, testCovar, ...)
}
\arguments{
\item{g}{The results of the groupLevelDifferencesWithCovariates test}

\item{testCovar}{The covariate tested by groupLevelDifferencesWithCovariates}

\item{donorListOne}{the donors in group one.}

\item{covarsDF}{The first column is named IID and contains the donor IDs.  Columns after the first contain covariates.}
}
\description{
A plot of effect sizes
}
