% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{fetchEqtlFiles}
\alias{fetchEqtlFiles}
\title{Fetch  a named list of eQTL files}
\usage{
fetchEqtlFiles(rootDir)
}
\arguments{
\item{rootDir}{root directory path to search for eQTL files}
}
\value{
named list of eQTL files with following names:
        eqtlResultsFile, indexSnpsFile, eigenMT.indexSnpsFile,
        eigenMT.resultsFile, expressionFile, peerExpressionFile
        genotypeMatrixFile, genotypeMatrixFileBgz, variantLocationsFile
}
\description{
Fetch  a named list of eQTL files
}
