% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderBiasedExpressionTest.R
\name{findGenderBiasedGenes}
\alias{findGenderBiasedGenes}
\title{Find genes that are biased by the gender of the donors.}
\usage{
findGenderBiasedGenes(
  expression_file_name,
  covariates_file_name,
  covar = "SEX",
  gene_location_file_name = NULL,
  numPermutations = 10000,
  outPDFExemplars = NULL,
  outPDFSummary = NULL,
  outFile = NULL
)
}
\arguments{
\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{covariates_file_name}{A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.  Contains a row with the donors' gender.}

\item{covar}{The row name that contains the donor copies of the X chromosome in the covariates file.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.  Optional.}

\item{numPermutations}{How many permutations should be run to generate permuted p-values}

\item{outPDFExemplars}{Output each gene as a plot}

\item{outPDFSummary}{Output the distribution of p-vals as a set of plots.}

\item{outFile}{The report of each gene tested and the p-values and effect sizes.}
}
\description{
Runs linear regression of gender (encoded as 1 for male, 2 for female - ie: copies of the X chromosome) vs expression of the gene.
Also runs permutations of the donor gender to generate a null, and permuted results by q-value.
}
