% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderBiasedExpressionTest.R
\name{findGenderBiasedGenesFromBaseFiles}
\alias{findGenderBiasedGenesFromBaseFiles}
\title{A convienience method for findGenderBiasedGenes}
\usage{
findGenderBiasedGenesFromBaseFiles(
  base.dir,
  base.name,
  out.dir = base.dir,
  covar = "SEX",
  numPermutations = 100
)
}
\arguments{
\item{base.dir}{The directory where the input files are located.}

\item{base.name}{The prefix name of the input files (usually an experiment name)}

\item{out.dir}{Where should the output files be generated?}

\item{covar}{The row name that contains the donor copies of the X chromosome in the covariates file.}

\item{numPermutations}{How many permutations should be run to generate permuted p-values}
}
\description{
A convienience method for findGenderBiasedGenes
}
