% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateGeneCorrelation.R
\name{findGeneCovariateCorrelation}
\alias{findGeneCovariateCorrelation}
\title{Find genes with biased expression based on one or more covariates.}
\usage{
findGeneCovariateCorrelation(
  expression_file_name,
  covariates_file_name,
  gene_location_file_name = NULL,
  numPermutations = 10000,
  numThreads = 1,
  outFile = NULL,
  filterCovarName = "SEX",
  filterCovarValue = NULL,
  contigGroupsFile = NULL,
  renormAutosomeExpression = F,
  covariatesToProcess = NULL,
  random.seed = 1234
)
}
\arguments{
\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{covariates_file_name}{A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.  Contains a row with the donors' gender.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.  Optional.}

\item{numPermutations}{How many permutations should be run to generate permuted p-values}

\item{numThreads}{The number of threads to run processing on.  Defaults to 1.}

\item{outFile}{The report of each gene tested and the p-values and effect sizes.}

\item{filterCovarName}{Filter donors by this covariate name to the subset that have the filterCovarValue}

\item{filterCovarValue}{Filter donors by this covariate name to the subset that have the filterCovarValue}

\item{contigGroupsFile}{A YAML file containing each contig name and groups it belongs to.  In this case the contigs should belong to the groups autosome and chrX.  If provided, renormalizes autosome expression data to not include the x chromosomne.}

\item{renormAutosomeExpression}{If a contigGroupsFile and this is true, then renormalize expression of the autosome to the sum of the autosome genes (instead of all genes)}

\item{covariatesToProcess}{If not null, filter covariates input to this list of IDs.}

\item{random.seed}{A random seed to ensure reproducibility across runs of the same data.}
}
\description{
Runs linear regression on all covariate terms linearly combined to explain the expression of a gene.
Also runs permutations of the donor label to generate a null, and permuted results by q-value.
}
