% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqltGwasSimpleColocation.R
\name{findGwasHits}
\alias{findGwasHits}
\title{For all GWAS hits above some pvalue threshold, find GWAS SNPs near or within a gene of interest.}
\usage{
findGwasHits(
  gwasFile,
  permutedQTLFile,
  unfilteredQTLFile,
  outFile,
  gwasHitThreshold = 5e-08,
  eQTLEmpiricThreshold = 1e-05,
  windowSize = 10000
)
}
\arguments{
\item{gwasFile}{A file containing GWAS SNP Hits}

\item{permutedQTLFile}{A file containing permuted eQTL hits}

\item{unfilteredQTLFile}{A file containing all unfiltered eQTL SNP tests.}

\item{outFile}{Output file.}

\item{gwasHitThreshold}{The threshold pvalue to select a GWAS SNP as a hit.}

\item{eQTLEmpiricThreshold}{The threshold pvalue to select an eQTL SNP as a hit.}

\item{windowSize}{How far around genes to search for an eQTL hit}
}
\value{
A list of genes with both a GWAS and EQTL hit.
}
\description{
Take the GWAS SNP hit and look up all nearby genes within some window size.
 Output the maximum qvalue gene within this region, as well as the empiric p-value for the SNP if tested.
 Order results by q-values.
}
