% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSecondaryEqtlEffect.R
\name{findSecondaryEqtlEffects}
\alias{findSecondaryEqtlEffects}
\title{Search for secondary eQTL effects at eGenes}
\usage{
findSecondaryEqtlEffects(
  indexSNPFile,
  snpByGeneResultFile,
  geneLocationFile,
  expressionFile,
  variantLocationFile,
  genotypeFile,
  outFile,
  outFileEigenMT,
  outPDF = NULL,
  fdrThreshold = 0.1,
  qvalueThreshold = 0.05,
  cisDist = 1e+06,
  dropSeqToolsLocation = "/broad/mccarroll/software/dropseq/priv",
  pythonPath = "/broad/software/free/Linux/redhat_7_x86_64/pkgs/python_3.9.2/bin/python"
)
}
\arguments{
\item{indexSNPFile}{A file containing the index SNP for each gene.  This contains the columns: SNP, gene, gene_start_pos, gene_end_pos, TESTS, qvalue.}

\item{snpByGeneResultFile}{A file containing all SNP by gene interactions for the original eQTL scan.}

\item{geneLocationFile}{The locations of genes in the scan.  This must have 4 columns in the header: geneid, chr, s1, s2.}

\item{expressionFile}{The expression values of the genes.  A matrix with an id column, followed by 1 column per donor, 1 row per gene.}

\item{variantLocationFile}{The location of the SNPs in the scan.  This must have 3 columns in the header: snp, chr, pos}

\item{genotypeFile}{The matrix of genotypes for the SNPs in the scan.  A matrix with an id column, followed by 1 column per donor, 1 row per SNP.
Genotypes are encoded as the number of copies of the alternate allele.}

\item{outFile}{Location to write the outfile to.  This contains all the new SNP x Gene interactions.}

\item{outFileEigenMT}{Location to write the final eigenMT result to.  This contains both the original index SNP results and new secondary effects
after correction by eigenMT.  There may be more than one SNP listed per gene, on seperate lines.}

\item{outPDF}{An optional file to write plots to.}

\item{fdrThreshold}{A FDR threshold to include secondary effects into the eigenMT testing.  Setting this to a high value will
result in all genes having a secondary effect included, which will increase multiple hypothesis testing burden.  Suggested value of 0.05.
Not all SNPs with an FDR threshold estimated <= fdrThreshold will have a final permuted p-value or q-value less than the threshold.}

\item{qvalueThreshold}{This threshold selects the set of eGenes to scan for secondary effects.  Suggested value 0.05.}

\item{cisDist}{What window size should be searched for secondary effects.  Suggested to use the same window size as the original search.}

\item{dropSeqToolsLocation}{Location where dropseq tools are installed.}

\item{pythonPath}{Path to the python executable.  This requires python 3 to run.  If python is on your path, this can be "python".}
}
\description{
Given a set of eGenes, test to see if there are additional eQTL effects from SNPs that
are not highly correlated with the index SNP.
}
\details{
This runs a linear regression via MatrixEQTL at each eGENE against every SNP in the window
except for the original index SNP.  The formula for each SNP in the window is expression ~ SNP + INDEX_SNP

Given those results, an estimate of the FDR is calculated for secondary eQTLs based on the number of tests generated
by the original eigenMT analysis of the window.  New SNPs with an FDR are included if they are below fdrThreshold.  EigenMT
is then re-run to include both primary and secondary effects.

If the outPDF argument is not null, 2 pages of plots per eGene will be generated.  The first page has the
region plot for the index SNP on the top, and the region plot after the index SNP has been accounted for on the bottom.

 The second page contains plots of the effect of the primary and secondary effects on expression of the gene, their
 individual and combined betas and how correlated the genotypes of the two SNPs are (R2).
}
