% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{fixIndels}
\alias{fixIndels}
\title{Find SNPs are indels in the gwas and eQTL data set and give them the same ref allele / direction of effect.}
\usage{
fixIndels(gwas, eQTL)
}
\arguments{
\item{gwas}{The GWAS data frame}

\item{eQTL}{The eQTL data frame}
}
\value{
The GWAS data frame, modified so that the alleles and direction of effect match the eQTL data
}
\description{
Finds SNPs that have D or I as one of their alleles in the GWAS data set and finds if the eQTL has a similar deletion
or insertion by looking at the number of bases in the allele.  Sets the reference/alt allele to match in the two data sets,
and flips alleles/odds ratios in the GWAS result as neccesary.
}
