% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{flipGwasAlleles}
\alias{flipGwasAlleles}
\title{Exchange the reference and alt as well as the direction of the odds ratio at the given index.}
\usage{
flipGwasAlleles(gwas, idxAllelesFlip)
}
\arguments{
\item{gwas}{The GWAS dataframe}

\item{idxAllelesFlip}{A vector of positions at which the alleles should be flipped.}
}
\value{
The GWAS data frame with the appropriate alleles/odds ratios flipped.
}
\description{
Flip the GWAS results alleles at the requested index.
Ths also flips the odds ratio (transform by 1/OR).
}
