% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{flipRefAltAlleles}
\alias{flipRefAltAlleles}
\title{Align the reference and alternate alleles of two sets of eQTL results so they match up}
\usage{
flipRefAltAlleles(df, sites, verbose = T)
}
\arguments{
\item{df}{A set of eQTL results where some alleles may need to be flipped.}

\item{sites}{A set of eQTL results that act as a reference for what allele should be considered
the reference and alternate at a particular SNP.}

\item{verbose}{If true print extra debug info on the number of alleles flipped.}
}
\value{
The df dataframe with the ref/alt alleles oriented to match the sites dataframe.
SNPs in the df that don't have the same alleles as the sites SNPs are filtered.  Flipped SNPs have their
beta multiplied by -1 as well to reverse the direction of effect of the eQTL.
}
\description{
Normally one would use the permuted data for the sites, and the unpermuted data for the df.
}
