% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{geneSnpPairAdaptivePermutation}
\alias{geneSnpPairAdaptivePermutation}
\title{Run adaptive permutation on one SNP Gene (or phenotype) pair.}
\usage{
geneSnpPairAdaptivePermutation(
  g,
  e,
  numPermutations = 1e+08,
  minNumPassingPermuations = 15,
  batchSize = 1e+06,
  startOOM = 5
)
}
\arguments{
\item{g}{The genotypes to test as the independent variable}

\item{e}{The expression value (or other response like a phenotype) to test.}

\item{numPermutations}{The maximum number of permutations}

\item{minNumPassingPermuations}{If an order of magnitude of permutations exceeed this number, stop permutation}

\item{batchSize}{how many permutations to run in a block.  Larger numbers use more memory}
}
\value{
data frame of results.
}
\description{
Run adaptive permutation on one SNP Gene (or phenotype) pair.
}
