% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{geneSNPPairPermutation}
\alias{geneSNPPairPermutation}
\title{Run the requested number of permutations distributed across batches}
\usage{
geneSNPPairPermutation(g, e, nPermutations, batchSize = 1e+06)
}
\value{
Report the number of trials and successees.
}
\description{
Run the requested number of permutations distributed across batches
}
