% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{geneSNPPairPermutations}
\alias{geneSNPPairPermutations}
\title{Run adaptive permutation teeting on a list of SNPs with a single phenotype}
\usage{
geneSNPPairPermutations(
  snpNames,
  empiricResults,
  phenotype,
  snps,
  numPermutations,
  minNumPassingPermuations = 15,
  batchSize = 1e+06
)
}
\arguments{
\item{snpNames}{The names of the SNPs to test.}

\item{empiricResults}{The partially permuted results from matrix eQTL}

\item{phenotype}{A numeric vector of phenotype or expression data to test against all SNP genotypes}

\item{snps}{The SNP genotype matrix.}

\item{numPermutations}{The maximum number of permutations that should be run}

\item{minNumPassingPermuations}{The number of permutation results > empiric result required to stop permutation before reaeching numPermutations}

\item{batchSize}{how many permutations should be run each iteration.  This defines the tradeoff of memory vs CPU time.}
}
\description{
This permutes a single genotype / phenotype pair.  The pair is permuted a maximum of
numPermutations times.  Permutation can end early if the number of permuted results that exceed
the empiric p-value is >= minNumPassingPermuations, at which point the permuted pvalue is
considered to be well estimated.  This can significantly short circuit permutation of a single SNP.
}
