% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{getAlleleFrequency}
\alias{getAlleleFrequency}
\title{Compute the minor allele frequency for a set of SNPs}
\usage{
getAlleleFrequency(snps)
}
\arguments{
\item{snps}{A set of SNP genotypes. SNPs can either be in a SlicedData object or a data.frame.}
}
\value{
A vector of allele frequencies, names set to the SNP IDs.
}
\description{
Calculates the frequency of the less common allele per SNP.
}
