% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateGeneCorrelation.R, R/eQTLUtils.R
\name{getContigGroups}
\alias{getContigGroups}
\title{Get the chromosomes that are part of the autosomes/X/Y/MT groups}
\usage{
getContigGroups(
  contigGroupsFile = NULL,
  autosomes = NULL,
  chrX = NULL,
  chrY = NULL,
  chrMT = NULL
)

getContigGroups(
  contigGroupsFile = NULL,
  autosomes = NULL,
  chrX = NULL,
  chrY = NULL,
  chrMT = NULL
)
}
\arguments{
\item{contigGroupsFile}{A yaml file containing the contig groups}

\item{autosomes}{The autosome contigs if not using the contigGroupsFile}

\item{chrX}{The X contigs if not using the contigGroupsFile}

\item{chrY}{The Y contigs if not using the contigGroupsFile}

\item{chrMT}{The MT contigs if not using the contigGroupsFile}
}
\value{
A list of the contigs on the autosome/X/Y/MT

A list of the contigs on the autosome/X/Y/MT
}
\description{
Get the chromosomes that are part of the autosomes/X/Y/MT groups

Get the chromosomes that are part of the autosomes/X/Y/MT groups
}
