% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{getCovariateTable}
\alias{getCovariateTable}
\title{Generate a flat contingency table for the covariates of interest}
\usage{
getCovariateTable(
  covariatesFile,
  covariatesOfInterest = c("SEX", "CELL_TYPE"),
  metaCellFile = NULL,
  rejectedDonorFile = NULL
)
}
\arguments{
\item{covariatesFile}{A file containing covariates.  The first column is named IID and contains the donor IDs.  Columns after the first contain covariates.
Each row contains the donor ID, followed by the values of the covariates.  If null, then no covariates will be accounted for in the analysis, and
the method will be unable to partition donors into groups by a covariate.}

\item{covariatesOfInterest}{A list of covariates to include from the covariates file in the contingency table}

\item{metaCellFile}{Extract expression data from a metacell file.  For each cell, this normalizes expression to sum to 100,000 across all genes.}

\item{rejectedDonorFile}{A file containing a list of donors that should be excluded from the analysis. Single column with no header.(Optional)}
}
\value{
a flat contingency table
}
\description{
Generate a flat contingency table for the covariates of interest
}
\seealso{
stats::ftable
}
