% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneSNPPositions.R
\name{getDistanceToGene}
\alias{getDistanceToGene}
\title{Add distance from SNP to gene in eQTL data.}
\usage{
getDistanceToGene(eQTLData, snps_location_file_name, gene_location_file_name)
}
\arguments{
\item{eQTLData}{data frame that has a SNP column for the SNP name and a gene column for the gene name}

\item{snps_location_file_name}{file in the standard eQTL format containing the columns snp,chr,pos that encode the snp name, chromosome, and position.}

\item{gene_location_file_name}{file in the standard eQTL format containing the columns geneid,chr,s1,s2 that encode the snp name, chromosome, and start position, end position.}
}
\value{
the eQTLData data frame annotated with the gene start/end positon, snp position, and distance to gene.
}
\description{
In an eQTL data frame, calculate the distance from the SNP to the gene.  SNPs found 5' of the gene have a negative distance, SNPs found 3' have a positive distance.  SNPs inside the gene are distance 0.
}
