% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{getDonorListsFromCovariates}
\alias{getDonorListsFromCovariates}
\title{For a given covariate in a covariate data frame that splits the donors into two groups, extract the donor IDs for each group.}
\usage{
getDonorListsFromCovariates(covarsDF, testCovar = NULL)
}
\arguments{
\item{covarsDF}{The first column is named IID and contains the donor IDs.  Columns after the first contain covariates.
To successfully partition the donor IDs into two groups, this column must contain 2 unique values or an error is returned.}

\item{testCovar}{The covariate name to group donors with.}
}
\value{
A list with 2 vectors of donor IDs, 1 per group.  If testCovar is null the returned list is NULL.
}
\description{
For a given covariate in a covariate data frame that splits the donors into two groups, extract the donor IDs for each group.
}
