% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{getEQTLResults}
\alias{getEQTLResults}
\title{Read the eQTL file.  Restrict to the best gene for each SNP if required}
\usage{
getEQTLResults(base.dir, base.name)
}
\arguments{
\item{eQTLFile}{The eQTL file to read in.}

\item{getBestGenePerSNP}{Should the set of SNPs be filtered to the best gene per SNP (each SNP appears once),
or should a SNP be allowed to appear for each gene it has been tested in?}
}
\value{
The dataframe of eQTL results.  Chromosome/position and ref/alt alleles are parsed out of the SNP name.  Chromosome X is set to chromosome 23.
}
\description{
Read the eQTL file.  Restrict to the best gene for each SNP if required
}
