% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_Example.R
\name{getErrorCovariance}
\alias{getErrorCovariance}
\title{Parse in the error covariance matrix if set, or return an empty numeric vector as expected by eQTL Matrix.}
\usage{
getErrorCovariance(errorCovarianceMatrixFile = NULL)
}
\arguments{
\item{errorCovarianceMatrixFile}{A tab seperated square matrix of donor kinship scores (multipled by 2), with row and column names.}
}
\value{
A square matrix of donor kinship scores if errorCovarianceMatrixFile is not null, otherwise a length 0 numeric vector
}
\description{
Parse in the error covariance matrix if set, or return an empty numeric vector as expected by eQTL Matrix.
}
