% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Extracts the sparse expression data matrix from an h5 seurat file.}
\usage{
getExpressionMatrix(
  h5SeuratFile,
  geneList = NULL,
  cellBarcodeList = NULL,
  filterExpression = F
)
}
\arguments{
\item{h5SeuratFile}{The .h5seurat file to load}

\item{geneList}{An optional list of genes.  Data will be subset to that list of genes}

\item{cellBarcodeList}{A list of cell barcodes.  If filterExpression is false, check that these are in the expected data.
If filterExpression is true, limit the expression data to the intersect of the cell barcodes in the list and expression data.}
}
\value{
a sparse matrix (dgCMatrix) of expression data.  Genes in rows, cells in columns.
}
\description{
Extracts the sparse expression data matrix from an h5 seurat file.
}
