% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{getFeatureList}
\alias{getFeatureList}
\title{Get a list of cell level features to use in data reduction and regression analysis}
\usage{
getFeatureList(
  fixedFeatures = c("SCALED_LOG_UMIS", "pct_mt"),
  factorName = NULL,
  usePeerFactors = FALSE,
  metricsDF,
  useAllLatentFactors = F,
  usageDF = NULL
)
}
\arguments{
\item{fixedFeatures}{The known cell level metadata}

\item{factorName}{A latent factor name to use in downstream analysis (optional)}

\item{usePeerFactors}{Should PEER factors be included in analysis?}

\item{metricsDF}{The cell level metadata that contains fixed features and PEER factors}

\item{useAllLatentFactors}{If true and usageDF is not null, extract all the latent factor names to include in analysis. (optional)}

\item{usageDF}{A matrix of latent factor by cell scores (optional)}
}
\value{
A list of features from the metricsDF and possibly usageDF that will be used by downstream analysis
}
\description{
Get a list of cell level features to use in data reduction and regression analysis
}
