% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{getFittedExpressionLinear}
\alias{getFittedExpressionLinear}
\title{Calculate the residuals of the fit of the expression data to the testCovariates.}
\usage{
getFittedExpressionLinear(
  expressionData,
  residualizeCovariates,
  covarsDF,
  method = c("linear", "robustLinear"),
  logExpression = F
)
}
\arguments{
\item{expressionData}{a data table of the expression data, with the rownames encoding the gene names, and the column names the donor names.}

\item{residualizeCovariates}{which covariates will be fit as the independent variables in the regression.  If this argument is of length 0, this function is a no-op.}

\item{covarsDF}{The data.frame or data table containing the covariates.  This contains the identifier in the first column (IID),
followed by one covariate per column, with the covariate name encoded in the column names.  Each row is a donor.}

\item{method}{Which linear regression method to use.  "linear" uses the default stats::lm function to fit the data, while
robustLinear uses MASS::rlm to fit the data and less heavily weigh outliers in the expression data.}

\item{logExpression}{If true, take the log of the expression data.  Any donor with expression 0 for a gene is set to the minimum non-zero value.}
}
\value{
A data.table with the same dimensions and encoding as the input expression data, with modified expression values.
}
\description{
For each gene, run a linear regression with expression as the dependent variable and the listed testCovariates as the independent variables.
Extract the residuals, then add the sum of the median expression to the residuals.
Emit a new matrix of expression with the same dimensions.
}
