% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{getFormattedGenomicRegion}
\alias{getFormattedGenomicRegion}
\title{Get the chr:start-end string from the eQTL/GWAS data frame.}
\usage{
getFormattedGenomicRegion(df)
}
\arguments{
\item{df}{A data frame containing the GWAS/eQTL combined data.}
}
\value{
The region for the data frame.  Always a single chromosome, with a start and end position.
}
\description{
Get the chr:start-end string from the eQTL/GWAS data frame.
}
