% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{getGroupLevelDifferencesFaster}
\alias{getGroupLevelDifferencesFaster}
\title{Compare two sets of expression data using the wilcoxon rank sum test}
\usage{
getGroupLevelDifferencesFaster(expressionData, donorListOne, donorListTwo)
}
\arguments{
\item{expressionData}{the matrix of expression data}

\item{donorListOne}{identifiers for group one}

\item{donorListTwo}{identifiers for group two}
}
\value{
a data frame of genes with p-values and BH corrected FDR.
}
\description{
see #https://www.researchgate.net/post/How_can_I_calculate_the_effect_size_for_Wilcoxon_signed_rank_test
for discussion of effect size used here.
}
