% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{getGwasEQTLOverlappingSNPs}
\alias{getGwasEQTLOverlappingSNPs}
\title{Find SNPs in both the GWAS and eQTL data sets and intersect}
\usage{
getGwasEQTLOverlappingSNPs(geneName, gwas, eQTL, verbose = F)
}
\arguments{
\item{geneName}{The gene to operate on}

\item{gwas}{The GWAS data frame}

\item{eQTL}{The eQTL data frame}

\item{verbose}{If true, write out the name of each gene as it is processed.}
}
\value{
A data frame for a gene containing the intersect of SNPs in the eQTL and GWAS data, with Z-scores calculated.
}
\description{
For the GWAS and eQTL data, find SNPs that are common to both sets.  Calculate the Z-score for the eQTL
data set based on the sign of the beta (>0 = +).
}
