% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{getGWasResults}
\alias{getGWasResults}
\title{Parse in GWAS results.}
\usage{
getGWasResults(gwasFile)
}
\arguments{
\item{gwasFile}{A GWAS file to parse.}
}
\value{
A dataframe containing the GWAS results.
}
\description{
If the chromosome name contains "chr", that is removed.
}
