% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{getOutputFileNameBase}
\alias{getOutputFileNameBase}
\title{Generate an output file based on the input file name and a specified output directory.}
\usage{
getOutputFileNameBase(inSNPFile, outDir)
}
\arguments{
\item{inSNPFile}{the input file name.  Used to generate a prefix name for all output files}

\item{outDir}{The output directory where files will be generated.}
}
\value{
The output file name prefix, containing the full path + basename of the output file.
}
\description{
Generate an output file based on the input file name and a specified output directory.
}
