% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{getQValuesAcrossExperiments}
\alias{getQValuesAcrossExperiments}
\title{Create the union of genes reported as significant in any eQTL experiment}
\usage{
getQValuesAcrossExperiments(
  experimentNames,
  permutedQTLFiles,
  addEffectSizes = F,
  qvalueThreshold = 0.05,
  geneDescriptionsFile = NULL,
  filterChromosomeList = NULL
)
}
\arguments{
\item{experimentNames}{The names of the experiment to appear in the columns of the outputs.}

\item{permutedQTLFiles}{A vector of eQTL result file that has gone through permutation (or other gene-level correction) and has a q-value.}

\item{addEffectSizes}{Emit the effect sizes of each significant eQTL.}

\item{qvalueThreshold}{A minimum q-value threshold for a gene from the permutedQTLFile to be included in the result set.}

\item{geneDescriptionsFile}{A file containing gene names in column 2 and gene descriptions in column 3.}

\item{filterChromosomeList}{remove chromosomes in this vector.  Useful when you want to filter out the X chromosome.}
}
\value{
A data frame with the gene, qvalues for experiments, and an optional gene description field.
}
\description{
Report the union of genes that were significant (<= qvalueThreshold) in any experiment, and the q-value in each experiment.
}
