% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{getSingleCellMetrics}
\alias{getSingleCellMetrics}
\title{Get cell level metrics, optionally filtered}
\usage{
getSingleCellMetrics(
  cellMetaDataFile,
  donorList = NULL,
  quantizationErrorThreshold = NULL
)
}
\arguments{
\item{cellMetaDataFile}{A matrix of cell level metadata}

\item{donorList}{An optional list of donors to subset the cell level metadata.}

\item{quantizationErrorThreshold}{If the cell level metadata contains the column QUANT_ERROR, the data has been previously compressed, and will be filtered to cells with values less than this score.}
}
\value{
A dataframe containing cell level metadata.
}
\description{
Get cell level metrics, optionally filtered
}
