% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{ggplotGenotypesVsExpression}
\alias{ggplotGenotypesVsExpression}
\title{Plot gene expression against a particular SNP using ggplot and ggbeeswarm}
\usage{
ggplotGenotypesVsExpression(
  gene,
  geneExpression,
  snp,
  rsid,
  genotype,
  files,
  outDir,
  as.obj = F,
  fullPath = NULL,
  experimentName = NULL
)
}
\arguments{
\item{gene}{name of gene being tested}

\item{snp}{string of the form {chrom}:{pos}:{ref}:{alt} specifying a SNP}

\item{rsid}{string giving RSID of SNP being tested (for showing in title/file name outputs)}

\item{genotype}{named integer vector representing the number of alt alleles at the given SNP for each donor}

\item{files}{named list of eQTl files returned from the function `fetchEqtlFiles`}

\item{outDir}{directory to write out any output files generated using naming scheme {paste0(outDir,'/',paste(gene,rsid,experimentName, sep="_")}}

\item{as.obj}{boolean flag that sets function to return ggplot object being plotted if set to TRUE. Will still write out if outDir or fullPath specified}

\item{fullPath}{path to output file used to overwrite default location relative to outDir (can leave outDir arg missing if this is specified)}

\item{experimentName}{if not specified, will use EXPERIMENT field of properties file}
}
\value{
only if as.obj == TRUE, ggplot object
}
\description{
Plot gene expression against a particular SNP using ggplot and ggbeeswarm
}
