% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{groupDifferencesByLinearRegression}
\alias{groupDifferencesByLinearRegression}
\title{Run a multivariate linear regression to explain gene expression by one or more covariates.}
\usage{
groupDifferencesByLinearRegression(
  expressionData,
  testCovariates,
  covarsDF,
  method = c("linear", "robustLinear")
)
}
\arguments{
\item{expressionData}{a data table of the expression data, with the rownames encoding the gene names, and the column names the donor names.}

\item{testCovariates}{which covariates will be fit as the independent variables in the regression}

\item{covarsDF}{The data.frame or data table containing the covariates.  This contains the identifier in the first column (IID),
followed by one covariate per column, with the covariate name encoded in the column names.  Each row is a donor.}

\item{method}{Which linear regression method to use.  "linear" uses the default stats::lm function to fit the data, while
robustLinear uses MASS::rlm to fit the data and less heavily weigh outliers in the expression data.  These weights
are then used to fit the standard linear regression with those outliers providing less influence on the regression.}
}
\value{
A data.table with the pvalue and beta per gene and testCovariate.
}
\description{
For each gene, run a linear regression with expression as the dependent variable and the listed testCovariates as the independent variables.
Extract the slope and p-value of each covariate's fit peer gene
}
