% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{groupLevelDifferencesWithCovariates}
\alias{groupLevelDifferencesWithCovariates}
\title{Given a set of covariates, remove the effects of covariates on gene expression before applying group level difference test.}
\usage{
groupLevelDifferencesWithCovariates(
  metaCellFile,
  rejectedDonorFile = NULL,
  covariatesFile = NULL,
  testCovar = NULL,
  residualizeCovariates = c(),
  contigGroupsFile,
  reducedGTFFile,
  donorListOne = NULL,
  donorListTwo = NULL,
  donorListOneFile = NULL,
  donorListTwoFile = NULL,
  autosomesOnly = F,
  normalizeAutosomeExpression = T,
  logExpression = T,
  pctGenesRetained = 50,
  quantileNormalize = F,
  method = c("linear", "robustLinear"),
  outFile = NULL,
  outPDF = NULL
)
}
\arguments{
\item{metaCellFile}{Extract expression data from a metacell file.  For each cell, this normalizes expression to sum to 100,000 across all genes.}

\item{rejectedDonorFile}{A file containing a list of donors that should be excluded from the analysis. Single column with no header.(Optional)}

\item{covariatesFile}{A file containing covariates.  The first column is named IID and contains the donor IDs.  Columns after the first contain covariates.
Each row contains the donor ID, followed by the values of the covariates.  If null, then no covariates will be accounted for in the analysis, and
the method will be unable to partition donors into groups by a covariate.}

\item{testCovar}{A covariate used to partition the data into two groups.  This covariate must split the data into exactly two groups,
for example 0/1, A/B, TRUE/FALSE, etc.  This is optional, and if null this method will fall back on the donorList parameters.}

\item{residualizeCovariates}{A list of covariates from the covariatesFile which will be accounted for in the regression and subtracted out of
the expression data.}

\item{contigGroupsFile}{A yaml file containing the contig groups.}

\item{reducedGTFFile}{The location of each gene.  Encodes the gene name, chromosome, start, end.}

\item{donorListOne}{A vector containing a list of donor IDs, which must be in the column names of the expressionFile.
This will select the donors in the first group.}

\item{donorListTwo}{An optional vector containing a list of donor IDs for the second group.  If NULL, assume all donors
in the expression file not in donorListOne are in this list.}

\item{donorListOneFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListOne.}

\item{donorListTwoFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListTwo.}

\item{autosomesOnly}{If true filter the metaCellMatrix to the autosome contigs before any normalization occurs}

\item{normalizeAutosomeExpression}{If true, normalize the expression of genes on the autosomes such that the total expression on the autosomes = numTranscriptsPerDonor.
Can not specify this option and autosomesOnly to both be true.}

\item{logExpression}{If true, use the log(expression) in the linear regression to remove the effects of confounding covariates.}

\item{pctGenesRetained}{Calculate the median number of transcripts per gene, and select this percent of the highly expressed genes to analyze.
This selection takes place after normalization to numTranscriptsPerDonor.}

\item{quantileNormalize}{If true data is quantile normalized.  This is probably a terrible idea.}

\item{method}{The linear regression model used to remove  effects of covariates on the expression data..  Model "linear" uses the standard regression model.
Model "robustLinear" uses MASS:rlm to find the best weights for each donor in the regression and coefficient for the model that is less influenced by outliers.}

\item{outFile}{writes the results data frame to this file.}

\item{outPDF}{If not null, emit plots to this file.}
}
\value{
A data frame containing the gene name, p-value for the group level test, and permuted p-value if requested.
}
\description{
This loads in a covariates file containing one more more covariates, then runs a linear regression using the expression as the dependent variable
and the selected covariates as independent variables.  From these regressions, the residuals are extracted and added to the median
expression of the gene.  This new gene expression accounts for the covariates, and is then used in the groupLevelDifferences test.
The gene level differences are calculated using the wilcoxon rank sum test to compare groups.
}
