% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{groupLevelDifferencesWithRobustRegression}
\alias{groupLevelDifferencesWithRobustRegression}
\title{Simultaneously test many covariates for their effects on expression}
\usage{
groupLevelDifferencesWithRobustRegression(
  metaCellFile,
  rejectedDonorFile = NULL,
  covariatesFile,
  testCovariates = NULL,
  reducedGTFFile,
  contigGroupsFile,
  donorListOne = NULL,
  donorListTwo = NULL,
  donorListOneFile = NULL,
  donorListTwoFile = NULL,
  autosomesOnly = T,
  normalizeAutosomeExpression = F,
  pctGenesRetained = 50,
  quantileNormalize = F,
  method = c("linear", "robustLinear"),
  outFile = NULL,
  outPDF = NULL
)
}
\arguments{
\item{metaCellFile}{Extract expression data from a metacell file.  For each cell, this normalizes expression to sum to 100,000 across all genes.}

\item{rejectedDonorFile}{A file containing a list of donors that should be excluded from the analysis. Single column with no header.(Optional)}

\item{covariatesFile}{A file containing covariates.  The first column is named IID and contains the donor IDs.  Columns after the first contain covariates.
Each row contains the donor ID, followed by the values of the covariates.  If null, then no covariates will be accounted for in the analysis, and
the method will be unable to partition donors into groups by a covariate.}

\item{testCovariates}{A list of covariates that will be tested for differential expression.}

\item{reducedGTFFile}{The location of each gene.  Encodes the gene name, chromosome, start, end.}

\item{contigGroupsFile}{A yaml file containing the contig groups.}

\item{donorListOne}{A vector containing a list of donor IDs, which must be in the column names of the expressionFile.
This will select the donors in the first group.}

\item{donorListTwo}{An optional vector containing a list of donor IDs for the second group.  If NULL, assume all donors
in the expression file not in donorListOne are in this list.}

\item{donorListOneFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListOne.}

\item{donorListTwoFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListTwo.}

\item{autosomesOnly}{If true filter the metaCellMatrix to the autosome contigs before any normalization occurs}

\item{normalizeAutosomeExpression}{If true, normalize the expression of genes on the autosomes such that the total expression on the autosomes = numTranscriptsPerDonor.
Can not specify this option and autosomesOnly to both be true.}

\item{pctGenesRetained}{Calculate the median number of transcripts per gene, and select this percent of the highly expressed genes to analyze.
This selection takes place after normalization to numTranscriptsPerDonor.}

\item{quantileNormalize}{If true data is quantile normalized.  This is probably a terrible idea.}

\item{method}{The linear regression model used find effects.  Model "linear" uses the standard regression model.
Model "robustLinear" uses MASS:rlm to find the best weights for each donor in the regression, then runs a standard
regression with lm using those weights.}

\item{outFile}{writes the results data frame to this file.}

\item{outPDF}{If not null, emit plots to this file.}
}
\value{
A data frame containing the gene name, beta, and p-value for each tested covariate
}
\description{
This tests each gene for differences in expression across multiple groups simultaniously.  Each gene's expression is the dependent variable
in a linear regression against all listed test covariates which are treated as independent variables.
}
\details{
Note that donor selection criteria for list one and two sub-select from the entire list, but there is no explicit donor grouping
in this operation, as many independent variables aree tested simultaniously.  If all donor options are NULL (the default) then all
donors in the intersect of the covariates and expression data are used.
}
