% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{indexGenotypeMatrix}
\alias{indexGenotypeMatrix}
\title{Reformat, bgzip, and index genotype matrix}
\usage{
indexGenotypeMatrix(genotypeMatrixFile, forceBgzip = F, forceIndex = F)
}
\arguments{
\item{genotypeMatrixFile}{file path for genotype matrix in txt.gz format}

\item{forceBgzip}{forces block-gzip compression of genotype matrix if true, otherwise uses existing file}

\item{forceIndex}{forces re-indexing of block-gzipped genotype matrix if true, otherwise uses existing file}
}
\value{
path to indexed bgzipped matrix
}
\description{
Writes out header as a single-line, tab-separated comment using the comment character '#'
}
