% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{interactionStdAnalysis}
\alias{interactionStdAnalysis}
\title{Generate eQTL interaction report}
\usage{
interactionStdAnalysis(
  interactionEqtlSummaryFile,
  outInteractionEqtlPDF,
  snpFile,
  h5SeuratFile,
  cellMetaDataFile,
  donorCovariatesFile,
  permutedPseudobulkEqtlResultsFile,
  usageFile,
  uMAPCoordinatesFile,
  qvalueThreshold = 0.05,
  factorName = c("FACTOR_2"),
  compressedImages = FALSE
)
}
\arguments{
\item{interactionEqtlSummaryFile}{The result of running singleCellEqtl}

\item{outInteractionEqtlPDF}{The location of the output PDF.}

\item{snpFile}{A SNP genotype matrix as generated by the eQTL pipeline}

\item{h5SeuratFile}{A file containing single cell expression data in h5 Seruat format}

\item{cellMetaDataFile}{A file containing a matrix of cell level metadata}

\item{donorCovariatesFile}{A file containing donor known covariates}

\item{permutedPseudobulkEqtlResultsFile}{The result file from the eQTL pipeline containg the permuted eQTL results generated by a pseudobulk analysis.}

\item{usageFile}{A file containing a matrix of latent factor scores}

\item{qvalueThreshold}{Filter snp/gene interactions in the permutedPseudobulkEqtlResultsFile}

\item{factorName}{Selects the factor(s) from the usageFile to analyze.  If null, regress all latent factors}

\item{compressedImages}{Use some semi-fancy ggplot stuff to compress each image as a PNG then write it to the PDF.
Causes no end of misery on UNIX, but works on OSX?}
}
\description{
Generate a qq plot of eGene : factor interactions, followed by a plot of each eGene eQTL effect.
}
