% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{intersectGwasAndLDBlock}
\alias{intersectGwasAndLDBlock}
\title{Builds LD blocks for the snps in the input data frame.}
\usage{
intersectGwasAndLDBlock(df, ldBlock, checkReversedAlleles = F)
}
\arguments{
\item{df}{A data frame containing the GWAS/eQTL combined data.}

\item{ldBlock}{A matrix of LD generated by bcftools for a region}

\item{checkReversedAlleles}{check the opposite allele orientation in the LD block.  When found, reverse the Z scores in the data frame.}
}
\value{
a list containing the GWAS/eQTL df and ldblock for the intersect of SNPs in both.
}
\description{
Because LDBlocks can only be querried out of a VCF by region, we need to select out the intersect of
variants that are in both the GWAS/eQTL data as well as the LD matrix.  The data is the ordered so
all data sets have their SNPs in the same order for CAVIAR/eCAVIAR.
}
