% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_PermutationTesting.R
\name{jankAssSortGenes}
\alias{jankAssSortGenes}
\title{Replicate matrix eQTL sorting by position.}
\usage{
jankAssSortGenes(gene, genepos, cisDist)
}
\arguments{
\item{gene}{A matrix of geneotypes}

\item{genepos}{A matrix of SNP locations.}

\item{cisDist}{The distance eQTL will search around each gene for SNPs.}
}
\value{
a list containing a sorted version of the gene and genepos objects handed in.
}
\description{
To avoid sorting data on every permutation iteration, sort the data in the strange way expected by matrix eQTL.
This feels a bit hacky, but does prevent sorting AND speeds up analysis.
}
