% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{loadExpressionData}
\alias{loadExpressionData}
\title{Load gene expression data}
\usage{
loadExpressionData(expression_file_name, useQuantileNormalization = F)
}
\arguments{
\item{expression_file_name}{The expression file}

\item{useQuantileNormalization}{Should the data be quantile normalized}
}
\value{
A SlicedData object holding the expression data.
}
\description{
Load gene expression data
}
