% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{loadGenotypeData}
\alias{loadGenotypeData}
\title{Load genotype data for eQTL analysis}
\usage{
loadGenotypeData(
  SNP_file_name,
  snps_location_file_name,
  minorAlleleFreqThreshold = NULL,
  snp_gene_map = NULL
)
}
\arguments{
\item{SNP_file_name}{genotypes => row: SNP; column: sample; value: {0,1,2,NA}}

\item{snps_location_file_name}{columns: snp, chr, pos, id}

\item{minorAlleleFreqThreshold}{Filter variants by minor allele frequency}

\item{snp_gene_map}{columns: snp, gene.  If present, this is joined to genotype data and
SNP location data.  Both genotype data and location data for a SNP are
duplicated for each gene that is to be considered with the SNP.  The SNP
name has the gene name appended to it to make MatrixEQTL work properly.}
}
\value{
list(snps, snpspos)
}
\description{
Load genotype data for eQTL analysis
}
