% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{makePerGeneSnpName}
\alias{makePerGeneSnpName}
\title{Convert SNP names so they are gene-specific}
\usage{
makePerGeneSnpName(snps, genes)
}
\arguments{
\item{snps}{vector of SNP names}

\item{genes}{vector of gene names}
}
\value{
vector of SNP-name:gene-name
}
\description{
Convert SNP names so they are gene-specific
}
