% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonParametricDifferentialExpression.R
\name{nonParametricDifferentialExpression}
\alias{nonParametricDifferentialExpression}
\title{Differential expression via wilcox test}
\usage{
nonParametricDifferentialExpression(
  metaCellFile,
  donorFile,
  alternative = "two.sided"
)
}
\arguments{
\item{metaCellFile}{A meta cell file, containing expression counts of donors, where each row contains data for a gene,
and each column the expression of a donor for that gene.}

\item{donorFile}{A file with no header containing the set of donors to test against all other donors in the data set.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
A data frame containing the gene names and thir wilcox.test p-value
}
\description{
For two sets of donors, test if the two sets of donors are part of the same continuous distribution.
Expression is converted from a continuous variable to a ranked representation.
If the given set of donor IDs is ranked higher or lower than the other donors, this will result in a
pvalue closer to 0.  In other words, the wilcoxon test determines if the expression values of the
selected set of donors shifted compared to all other donors.
}
