% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeExpressionByCovars.R
\name{normalizeExpressionByCovariates}
\alias{normalizeExpressionByCovariates}
\title{Remove the effect of covariates from the expression data}
\usage{
normalizeExpressionByCovariates(
  expressionFile,
  covariateFile,
  outExpressionFile = NULL
)
}
\arguments{
\item{expressionFile}{The input expression data set - columns: id column for gene followed by donor names.
rows: gene name followed by expression level per donor}

\item{covariateFile}{The input covariate data set - columns: id column for covariate followed by donor names.
rows: covariate name followed by covariate values per donor.}

\item{outExpressionFile}{The ouput file, in the same format as the input.}
}
\value{
If outExpressionFile is NULL, return the modified expression data frame.
}
\description{
Given the expression and covariate inputs, regress out the covariates from the expression data
and generate new expression data. This is done by running a linear regression of
expression ~ covars, extracting the residuals and then adding the mean expression to the residuals.
}
