% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLPlots.R
\name{plotDiversity}
\alias{plotDiversity}
\title{Plot diversity of UMIs/Donors in an eQTL experiment.}
\usage{
plotDiversity(metaCellFile, outPDF = NULL)
}
\arguments{
\item{metaCellFile}{A single file of metacells.}

\item{outPDF}{The PDF file to write to, or NULL.}
}
\description{
Plot diversity of UMIs/Donors in an eQTL experiment.
}
