% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLPlots.R
\name{plotDonorBalanceByAssignableUMIs}
\alias{plotDonorBalanceByAssignableUMIs}
\title{A stacked area plot by donor assignable UMIs.}
\usage{
plotDonorBalanceByAssignableUMIs(
  metaCellFiles,
  experimentNames,
  plotTitle = ""
)
}
\arguments{
\item{metaCellFiles}{A vector of metaCell files}

\item{experimentNames}{a vector of matching label names}

\item{plotTitle}{A plot title.}
}
\description{
A stacked area plot by donor assignable UMIs.
}
