% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneRegion.R
\name{ploteQTLRegion}
\alias{ploteQTLRegion}
\title{Plot the cis eQTL region for a single gene.}
\usage{
ploteQTLRegion(
  geneName,
  start,
  end,
  cisEQTLs,
  geneLoc,
  ratioSNPsToGenes = 0.6,
  indexSNP = NULL,
  mafColumn = NULL
)
}
\arguments{
\item{geneName}{the gene to plot, must be in the cisEQTLs data frame gene column.}

\item{start}{The start position to plot from}

\item{end}{The end position to plot to.}

\item{cisEQTLs}{is a data frame that includes the following columns:  gene, chr, pos, p-value.}

\item{geneLoc}{is a data frame of gene locations.  Has the following columns: geneid, chr, s1, s2.
s1 is the start of the gene, and s2 is the end.}

\item{ratioSNPsToGenes}{what fraction of the figure should be used by the SNP part of the plot?  [0-1].}

\item{indexSNP}{A SNP selected from the eQTL data (SNP column) as the best SNP for the region,
which will be highlighted in red on the SNP plot.  (Optional)}

\item{mafColumn}{if provided, will color eQTL snps by minor allele frequency found in this column}
}
\description{
This is a plot of the cis eQTL pvalues and the genes.
}
