% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderBiasedExpressionTest.R
\name{plotExemplarGender}
\alias{plotExemplarGender}
\title{Plot a single gene gender vs expression with fit.}
\usage{
plotExemplarGender(geneName, expData, result, sex)
}
\arguments{
\item{geneName}{The gene name}

\item{expData}{A SlicedData set of expression data}

\item{result}{The eQTL-type scan of all genes}

\item{sex}{A SlicedData set of the sex of the donors}
}
\description{
Plot a single gene gender vs expression with fit.
}
