% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLPlots.R
\name{plotGeneIntersection}
\alias{plotGeneIntersection}
\title{Plot Venn Diagrams of genes found in multiple experiments}
\usage{
plotGeneIntersection(inFiles, experimentNames, qvalueThreshold = 0.05)
}
\arguments{
\item{inFiles}{A list of eQTL result files that have a qvalue column.}

\item{experimentNames}{The names of the experiments for the plot}

\item{qvalueThreshold}{the maximum q-value genes that will be retained.}
}
\description{
Plot Venn Diagrams of genes found in multiple experiments
}
