% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLeveleQTLPlots.R
\name{plotGeneQTLsForMultipleExperiments}
\alias{plotGeneQTLsForMultipleExperiments}
\title{For multiple experiments, plot the eQTL results for the same gene same SNP across multiple experiments.}
\usage{
plotGeneQTLsForMultipleExperiments(
  gene_location_file_name,
  snps_location_file_name,
  expression_file_name,
  SNP_file_name,
  eqtl_result_file_name,
  eQTLPermutationResultFile,
  expNames,
  outPDF = NULL,
  qValueThreshold = 0.05,
  geneList = NULL,
  mode = c("union", "intersect", "disagree")
)
}
\arguments{
\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{eQTLPermutationResultFile}{The permuted eQTL result file}

\item{expNames}{the experiment names for the input files}

\item{outPDF}{The file location to output the PDF of SNP/Expression plots.}

\item{qValueThreshold}{Genes with a score <= this threshold will be emitted in the outIndexSNPs and outPDF.}

\item{geneList}{Override the qValueThreshold and plot these genes instead.}

\item{mode}{One of union/intersect/disagree.  Union picks all genes in any experiment.  Intersect picks genes in all experiments.  Disagree selects genes that are positive in one experiment and negative in another.}
}
\description{
Each input parameter is a list of files. Results are plotted for genes that are significant (q-value < threshold) in any experiment.
}
