% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLeveleQTLPlots.R
\name{plotGeneQTLsForMultipleExperimentsBaseNames}
\alias{plotGeneQTLsForMultipleExperimentsBaseNames}
\title{For multiple experiments, plot the eQTL results for the same gene same SNP across multiple experiments.}
\usage{
plotGeneQTLsForMultipleExperimentsBaseNames(
  dataDirList,
  basenames,
  expNames,
  outPDF = NULL,
  qValueThreshold = 0.05,
  geneList = NULL,
  mode = c("union", "intersect", "disagree")
)
}
\arguments{
\item{dataDirList}{The directory data resides in - this is a vector of entries.}

\item{basenames}{The prefix of all files for this data set - this is a vector of entries of the same length as dataDirList}

\item{expNames}{The experiment names for the data, which will be in the header of the plots.}

\item{outPDF}{Where to write a PDF output}

\item{qValueThreshold}{Genes with a score <= this threshold will be emitted in the outIndexSNPs and outPDF.}

\item{geneList}{Override the qValueThreshold and plot these genes instead.}

\item{mode}{One of union/intersect/disagree.  Union picks all genes in any experiment.  Intersect picks genes in all experiments.  Disagree selects genes that are positive in one experiment and negative in another.}
}
\description{
Each input parameter is a list of files. Results are plotted for genes that are significant (q-value < threshold) in any experiment.
}
