% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLeveleQTLPlots.R
\name{plotGeneQTLsFromBaseFileNames}
\alias{plotGeneQTLsFromBaseFileNames}
\title{Plot expression by genotype plots for eQTLs below some qvalue threshold.}
\usage{
plotGeneQTLsFromBaseFileNames(
  data.dir,
  base.name,
  eQTL.dir,
  eQTLPermutationResultFile = NULL,
  qValueThreshold = 0.05
)
}
\arguments{
\item{data.dir}{The directory data resides in.}

\item{base.name}{The prefix of all files for this data set.}

\item{eQTL.dir}{The directory the eQTL permuted results reside in.  Files will be written out here.}

\item{eQTLPermutationResultFile}{a set of results for the eQTLs that have been permuted and have qvalues.  If not supplied,
will attempt to infer the proper file name given the base name and eQTL.dir where it usually resides.}

\item{qValueThreshold}{the threshold at which eQTLs are selected.  Selects all eQTLs with <=threhold value.}
}
\value{
The results of adaptive permutation (ALL SNPs per gene.)
}
\description{
Plot expression by genotype plots for eQTLs below some qvalue threshold.
}
