% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneRegion.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes.}
\usage{
plotGenes(
  genes,
  s,
  e,
  highlightFeatureRegion = NULL,
  highlightColor = "red",
  plotXAxis = T
)
}
\arguments{
\item{genes}{a data frame of gene locations. Has the following columns: geneid, chr, s1, s2.
s1 is the start of the gene, and s2 is the end.}

\item{s}{the start of the region}

\item{e}{the end of the region}

\item{highlightFeatureRegion}{a data frame with two columns start and end, that defines regions to highlight a different color}

\item{highlightColor}{What color should the highlights be?}

\item{plotXAxis}{should the genomic locations with tick marks be plotted?}
}
\description{
Plot a set of gene locations as a series of lines.
}
