% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SexBiasedGeneExpression.R
\name{plotGenesXXDosage}
\alias{plotGenesXXDosage}
\title{Plots the SEX/FRACTION_X regression results.}
\usage{
plotGenesXXDosage(
  expression_file_name,
  covariates_file_name,
  gene_location_file_name,
  permutedResultsFile,
  contigGroupsFile,
  renormAutosomeExpression = T,
  pValueThreshold = 5e-08,
  outPDF = NULL
)
}
\arguments{
\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{covariates_file_name}{A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.  Contains a row with the donors' gender.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{permutedResultsFile}{The report of each gene tested and the p-values and effect sizes for the covariates SEX and FRACTION_X}

\item{contigGroupsFile}{A YAML file containing each contig name and groups it belongs to.  In this case the contigs should belong to the groups autosome and chrX.  If provided, renormalizes autosome expression data to not include the x chromosomne.}

\item{renormAutosomeExpression}{If a contigGroupsFile and this is true, then renormalize expression of the autosome to the sum of the autosome genes (instead of all genes)}

\item{pValueThreshold}{Plot all genes that have at a pvalue < this score in one category.}

\item{outPDF}{Output file for plots.}
}
\description{
Plots the SEX/FRACTION_X regression results.
}
