% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLeveleQTLPlots.R
\name{plotGenotypesVsExpression}
\alias{plotGenotypesVsExpression}
\title{Plot the genotype by expression regression analysis plot for a single SNP and gene.}
\usage{
plotGenotypesVsExpression(
  geneName,
  cisEQTLs,
  snpLoc,
  expressionMatrix,
  geneotypes,
  strTitlePrefix = "",
  indexSNP = NULL,
  ylim = NULL,
  eQTLResult = NULL
)
}
\arguments{
\item{geneName}{The name of the gene to plot}

\item{cisEQTLs}{The list of eQTL results (can be permuted or all hits.)}

\item{snpLoc}{A data frame of snp locations}

\item{expressionMatrix}{A data frame of expression data per gene/donor}

\item{geneotypes}{A matrix of genotypes (copies of alternate allele) per snp/donor}

\item{strTitlePrefix}{Put this at the front of the plot}

\item{indexSNP}{A SNP name to plot.  If not supplied, and you passed in the permuted hits, uses the best hit SNP for the gene.}

\item{ylim}{Limit the expression of the Y axis (optional)}

\item{eQTLResult}{All SNP by gene results for this experiment (optional).  If provided, get the empiric pvalue from this dataframe
if the SNP/Gene pair does not exist in the cisEQTLs result.  Useful when cisEQTLs is the permuted result and an index SNP is defined.}
}
\description{
Plot the genotype by expression regression analysis plot for a single SNP and gene.
}
