% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{plotGenotypesVsExpressionAcrossResults}
\alias{plotGenotypesVsExpressionAcrossResults}
\title{Given a vector of root directories and a queryFile, runs set of
SNP/Gene queries across eQTL results contained within each root directory}
\usage{
plotGenotypesVsExpressionAcrossResults(
  rootDirs,
  queryFile,
  outDir,
  usePeer = T
)
}
\arguments{
\item{rootDirs}{character vector of directories containing one set of eQTL results}

\item{queryFile}{A tab delimited file with 2 columns.  SNP: the rsID of the SNP of interest.}

\item{outDir}{output directory to write out files into. Files are written out}

\item{usePeer}{specifies whether or not to use peer-adjusted expression data for testing queries}
}
\description{
Given a vector of root directories and a queryFile, runs set of
SNP/Gene queries across eQTL results contained within each root directory
}
