% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneRegion.R
\name{plotGWASRegion}
\alias{plotGWASRegion}
\title{Plot a GWAS region}
\usage{
plotGWASRegion(
  geneName,
  gwas,
  chrName,
  start,
  end,
  geneLoc,
  ratioSNPsToGenes = 0.6,
  indexSNP = NULL,
  pvalColumn = "p"
)
}
\arguments{
\item{geneName}{the gene to plot.  Only affects the plot title.}

\item{gwas}{The data frame of gwas data to draw SNPs from.  expected columns chr, pos, and a pvalue column that is by default 'p'}

\item{chrName}{The name of the chromosome to draw data from.}

\item{start}{The start position to plot from}

\item{end}{The end position to plot to.}

\item{geneLoc}{A dataframe of gene locations.  Has 4 columns: geneid, chr, s1, s2.  S1 is the start, S2 is the end of the gene.}

\item{ratioSNPsToGenes}{what fraction of the figure should be used by the SNP part of the plot?  [0-1].}

\item{indexSNP}{A SNP selected from the GWAS data as the best SNP for the region.}

\item{pvalColumn}{The column name to look for the SNP pvalue in.}
}
\description{
Plots the SNP pvalues [-log10] as well as the genes in the region.
}
