% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{plotManyEqtls}
\alias{plotManyEqtls}
\title{Plot a group of eQTLs}
\usage{
plotManyEqtls(data, gene, snp, bTotal, qvalue = NULL, base_size = 11)
}
\arguments{
\item{data}{Data as generated by getExpressionBinnedByFactorScore}

\item{gene}{The name of the gene being plotted}

\item{snp}{The name of the SNP being plotted}

\item{bTotal}{The BTotal score per cell as generated by getExpressionBinnedByFactorScore}

\item{qvalue}{Qvalue of the eQTL (optional)}

\item{base_size}{The ggplot2 base size for the plot}
}
\value{
A plot_grid result
}
\description{
Plot a group of eQTLs
}
