% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{plotResults}
\alias{plotResults}
\title{Plot the CAVIAR/eCAVIAR fine mapping results on a region.}
\usage{
plotResults(outputBaseFileName, gene_location_file_name, outPDFFile = NULL)
}
\arguments{
\item{outputBaseFileName}{The prefix for all files for this region, including the full directory path.}

\item{gene_location_file_name}{File containing the coordinates of all genes in the eQTL study.}

\item{outPDFFile}{If not null, create a new PDF file and plot to it.}
}
\description{
Plot the CAVIAR/eCAVIAR fine mapping results on a region.
}
