% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareGWASData.R
\name{prepareGWASData}
\alias{prepareGWASData}
\title{Convert GWAS data to a standard format}
\usage{
prepareGWASData(
  inFile,
  outFile,
  imputationThreshold = 0.9,
  inSequenceDictionaryFile = NULL,
  outputSequenceDictionaryFile = NULL,
  liftOverChainFile = NULL,
  picardPath = NULL
)
}
\arguments{
\item{inFile}{The input GWAS file - this accepts a variety of standard formats.}

\item{outFile}{The GWAS data, reformatted, filtered, lifted over, and sorted for indexing.}

\item{imputationThreshold}{The minimum imputation threshold to retain a marker}

\item{inSequenceDictionaryFile}{The sequence dictionary file containing the contigs for input file.}

\item{outputSequenceDictionaryFile}{The sequence dictionary file containing the contigs for the output file}

\item{liftOverChainFile}{The chain file that maps position mapping from the old to new genome builds}

\item{picardPath}{Where the picard jar file is located.}
}
\description{
Convert GWAS data to a standard format that can be read in by LDScore, used in GWAS correlation analysis,
and is suitable to be indexed by Tabix.
}
\details{
The inSequenceDictionaryFile, outputSequenceDictionaryFile, liftOverChainFile, and picardPath can be left null
unless you desire to lift over your GWAS results to a different build.
}
