% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R, R/eQTLUtils.R
\name{quantileNormalizeExpression}
\alias{quantileNormalizeExpression}
\title{Quantile normalize expression data
The goal of this is to normally distribute expression data that may not be normally distributed.
See: https://en.wikipedia.org/wiki/Quantile_normalization}
\usage{
quantileNormalizeExpression(expressionData)

quantileNormalizeExpression(expressionData)
}
\arguments{
\item{expressionData}{The data.frame or data.table of expression data to normalize.  Rownames contain genes, column
names contain donor identifiers.  The values of the data.frame are all numeric.}
}
\value{
A data.table of expression data with the same number of rows and columns as the input data, but with quantile normalized expression values

A data.table of expression data with the same number of rows and columns as the input data, but with quantile normalized expression values
}
\description{
Quantile normalize expression data
The goal of this is to normally distribute expression data that may not be normally distributed.
See: https://en.wikipedia.org/wiki/Quantile_normalization

Quantile normalize expression data
The goal of this is to normally distribute expression data that may not be normally distributed.
See: https://en.wikipedia.org/wiki/Quantile_normalization
}
