% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAcrossEqtlResults.R
\name{queryTabixFile}
\alias{queryTabixFile}
\title{Quickly query a tabix-indexed bgzipped tsv file}
\usage{
queryTabixFile(tabixFile, chr, irange, delim = "\\t", comment = "#")
}
\arguments{
\item{tabixFile}{path to tabix-indexed bgzipped file}

\item{chr}{chromosome to query}

\item{irange}{object of class IRanges from Rsamtools defining positions to query}

\item{delim}{field delimiter of tabixFile defaults to tab}

\item{comment}{comment character used in tabixFile}
}
\value{
data frame with results of tabix query
}
\description{
Quickly query a tabix-indexed bgzipped tsv file
}
