% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePermutationResults.R
\name{readFile}
\alias{readFile}
\title{Read in a eQTL permutation result file.}
\usage{
readFile(x)
}
\arguments{
\item{x}{A single chromosome permutation result file.}
}
\value{
A data table containing the single chromosome permutation results.
An attribute gene_count is added that has the total number of genes tested.
}
\description{
Read in a eQTL permutation result file.
}
