% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{readGroupLevelDifferencesData}
\alias{readGroupLevelDifferencesData}
\title{Reads in the data neccesary to do group level tests}
\usage{
readGroupLevelDifferencesData(
  metaCellFile,
  rejectedDonorFile = NULL,
  donorListOne,
  donorListTwo = NULL,
  donorListOneFile = NULL,
  donorListTwoFile = NULL,
  reducedGTFFile,
  contigGroupsFile,
  autosomesOnly = T,
  normalizeAutosomeExpression = F,
  pctGenesRetained = 50,
  validateDonorsInExpData = T,
  filterExpressionToDonorList = T
)
}
\arguments{
\item{metaCellFile}{Extract expression data from a metacell file.  For each cell, this normalizes expression to sum to 100,000 across all genes.}

\item{rejectedDonorFile}{A file containing a list of donors that should be excluded from the analysis. Single column with no header.(Optional)}

\item{donorListOne}{A vector containing a list of donor IDs, which must be in the column names of the expressionFile.
This will select the donors in the first group.}

\item{donorListTwo}{An optional vector containing a list of donor IDs for the second group.  If NULL, assume all donors
in the expression file not in donorListOne are in this list.}

\item{donorListOneFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListOne.}

\item{donorListTwoFile}{If set, reads in the donorListOne from this file.  Overrides settings to donorListTwo.}

\item{reducedGTFFile}{a data frame with column names gene_name, chr, start, end}

\item{contigGroupsFile}{A yaml file containing the contig groups.}

\item{autosomesOnly}{If true filter the metaCellMatrix to the autosome contigs before any normalization occurs}

\item{normalizeAutosomeExpression}{If true, normalize the expression of genes on the autosomes such that the total expression on the autosomes = numTranscriptsPerDonor.
Can not specify this option and autosomesOnly to both be true.}

\item{pctGenesRetained}{Calculate the median number of transcripts per gene, and select this percent of the highly expressed genes to analyze.
This selection takes place after normalization to numTranscriptsPerDonor.}

\item{validateDonorsInExpData}{Test to see if all donors in the donor lists are in the metaCellFile}

\item{filterExpressionToDonorList}{If true, subsets the donors in the metaCellFile to the donors in the union of the donor lists.}
}
\value{
A list with 4 elements: 1) A data frame of expression data 2) a data frame of gene locations
3) A vector donor IDs in group 1. 4) A vector donor IDs in group 2.
}
\description{
Reads in the data neccesary to do group level tests
}
