% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePhenotypePlots.R
\name{readGWASInput}
\alias{readGWASInput}
\title{Read the GWAS result from permutation testing and annotate / clean up the input for analysis}
\usage{
readGWASInput(gwasResult, variantLocationFile)
}
\description{
Read the GWAS result from permutation testing and annotate / clean up the input for analysis
}
