% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{readMetaCellData}
\alias{readMetaCellData}
\title{Read in the meta cells file and normalize to look like the expression data.}
\usage{
readMetaCellData(
  metaCellFile,
  rejectedDonorFile,
  reducedGTF,
  contigGroupsFile,
  numTranscriptsPerDonor = 1e+05,
  autosomesOnly = T,
  normalizeAutosomeExpression = F,
  pctGenesRetained = 50
)
}
\arguments{
\item{metaCellFile}{Extract expression data from a metacell file.  For each cell, this normalizes expression to sum to 100,000 across all genes.  The top 50%
of genes are selected for downstream analysis.}

\item{rejectedDonorFile}{A file containing a list of donors that should be excluded from the analysis. Single column with no header.(Optional)}

\item{reducedGTF}{a data frame with column names gene_name, chr, start, end}

\item{contigGroupsFile}{A yaml file containing the contig groups.}

\item{numTranscriptsPerDonor}{Normalize expression data such that each donor (column of the matrix) sums to this many transcripts}

\item{autosomesOnly}{If true filter the metaCellMatrix to the autosome contigs before any normalization occurs}

\item{normalizeAutosomeExpression}{If true, normalize the expression of genes on the autosomes such that the total expression on the autosomes = numTranscriptsPerDonor.
Can not specify this option and autosomesOnly to both be true.}

\item{pctGenesRetained}{Calculate the median number of transcripts per gene, and select this percent of the highly expressed genes to analyze.
This selection takes place after normalization to numTranscriptsPerDonor.}
}
\value{
a data frame of expression data, where each column is a donor and each row is a gene.
}
\description{
Convert the meta cell counts data into normalized expression data.
}
