% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelDifferences.R
\name{readReducedGTF}
\alias{readReducedGTF}
\title{Reads in a reduced GTF file}
\usage{
readReducedGTF(reducedGTFFile)
}
\arguments{
\item{reducedGTFFile}{The location of each gene.  Encodes the gene name, chromosome, start, end.}
}
\value{
a data frame with column names gene_name, chr, start, end
}
\description{
Reads in a reduced GTF file
}
