% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{reduceData}
\alias{reduceData}
\title{Reduce the single cell representation to a less granular data set to reduce computational burden}
\usage{
reduceData(
  numClustersPerDonor = NULL,
  featureName,
  featureList,
  metricsDF,
  expData,
  usageDF,
  reduceDataPDF = NULL
)
}
\arguments{
\item{numClustersPerDonor}{The number of cells to emit for each donor.  These cells are the "average" cells for each cluster.}

\item{featureName}{A column in the metricsDF.  This feature will be plotted in the output PDF,
and a measurement of the error of each cell to it's median will be calculated.  This only affects plotting, not the data outputs.
This is typically a latent factor of interest.  This parameter only affects the error calculation.
This feature should also be in the featureList (and will be added if it is not.)}

\item{featureList}{A vector of features that will used to calculate the distances between cells.
This list typically includes latent factor(s) of interest and known covariates.  All features
must be numeric.}

\item{metricsDF}{A data frame containing the features for each cell barcode.}

\item{expData}{A matrix containing expression data for each cell - cells in rows, genes in columns.}

\item{usageDF}{A data frame containing the latent factor usage per cell.  Each row repesents a single cell,
columns represent each factor score.  Cell IDs are stored in the rownames of the dataframe.}

\item{reduceDataPDF}{An optional PDF to emit plots to.  Each plot contains a summary of how the cells of a donor
were reduced to a set of exemplar cells.}
}
\value{
A list containing the modified expression and metrics dataframes.  These are drop-in replacements for the
original expData and metricsDF data that was passed in.
}
\description{
This leverages vector quantization to split each donor's cells into N groups where cells in a group have
similar scores for known covariate information.  In effect, this bins each donor's total cells into N bins of data representing
the cluster center for the group
}
\details{
# The expression data is modified to reflect the median of the cells in the group.  This preserves the counts nature of the data.
The expression data is modified to reflect the sum of cells in the group.  The number of total transcripts is recalculated to the sum of the cells
in the group, and the SCALED_LOG_UMIS feature is recalculated for the new exemplar cells.

The cell barcode names for the metricsDF and expression data are changed to reflect the "bins" of donor expression.
}
