% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{reduceDataCmdLine}
\alias{reduceDataCmdLine}
\title{Reduce the single cell representation to a less granular data set to reduce computational burden}
\usage{
reduceDataCmdLine(
  snpFile,
  h5SeuratFile,
  cellMetaDataFile,
  donorCovariatesFile,
  usageFile,
  factorName,
  useAllLatentFactors = F,
  usePeerFactors = F,
  numClustersPerDonor = 20,
  permutedPseudobulkEqtlResultsFile,
  qvalueThreshold = 1,
  eQTLGeneList = NULL,
  chromosome = NULL,
  reduceDataPDF = NULL,
  outCellFeaturesFile,
  outExpressionFile,
  outUsageFile
)
}
\arguments{
\item{numClustersPerDonor}{The number of cells to emit for each donor.  These cells are the "average" cells for each cluster.}

\item{reduceDataPDF}{An optional PDF to emit plots to.  Each plot contains a summary of how the cells of a donor
were reduced to a set of exemplar cells.}

\item{outCellFeaturesFile}{A cell level metrics file of the compressed cells}

\item{outExpressionFile}{A seurat h5 file of expression for the compressed cells}

\item{outUsageFile}{A latent factor matrix containing the selected factor for the compressed cells}
}
\description{
This is the command line version of reduceData that reads in inputs from files and emits flat files
containing the new expression data, cell features, and factor usage generated by data collapse.
}
\details{
These new files can then be 1-1 replacements for the original inputs and used in analysis.

The permutedPseudobulkEqtlResultsFile provides a pre-filtered set of genes to operate on.  This
can additionally be filtered to a specific chromosome, gene list, or qvalue threshold.
}
