% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLUtils.R
\name{removeCovarEffectOnExpression}
\alias{removeCovarEffectOnExpression}
\title{Get the residual expression after removing the effects of one or more covariates}
\usage{
removeCovarEffectOnExpression(covarNames, expData, covars)
}
\arguments{
\item{covarNames}{The covariates to subtract from the expression data}

\item{expData}{The expression data}

\item{covars}{The covariate data}
}
\value{
A data frame in the same format as the input expression data, with the effects of the covariate removed.
}
\description{
Fit a linear regression to explain how much of expression variance is explained by a covariate
IE: lm (expression ~ covariate)
Extract the residuals from the regression and add the mean expression.
}
