% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLUtils.R
\name{renormalizeAutosomeExpression}
\alias{renormalizeAutosomeExpression}
\title{Renormalize expression data from gene/sum(all genes) to gene/sum(autosome genes).}
\usage{
renormalizeAutosomeExpression(
  expData,
  geneLoc = NULL,
  reducedGTF = NULL,
  contigGroupsFile = NULL
)
}
\arguments{
\item{expData}{The expression data.  This must be a data frame, NOT a data table.}

\item{geneLoc}{The location of each gene.  Encodes the gene name, chromosome, start, end. (optional, can use this or reducedGTF)}

\item{reducedGTF}{The location of each gene.}

\item{contigGroupsFile}{A yaml file containing the contig groups}
}
\value{
The expression data after normalization
}
\description{
Renormalize expression of the autosomes to sum to 1, instead of including the X/Y/MT chromosomes.
}
