% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_PermutationTesting.R
\name{runAdaptiveMultiStepPermutations}
\alias{runAdaptiveMultiStepPermutations}
\title{Multi-step permutation analysis}
\usage{
runAdaptiveMultiStepPermutations(
  snps_location_file_name,
  SNP_file_name,
  expression_file_name,
  gene_location_file_name,
  covariates_file_name = NULL,
  eQTLInputFile,
  permutationFileName,
  eQTLOutputFullPermutationFile,
  snpReportFile,
  maxNumIter = 1000,
  cisDist = 1e+05,
  minNumPassingPermuations = 5,
  useModel = modelLINEAR,
  minorAlleleFreqThreshold = NULL,
  useQuantileNormalization = F,
  errorCovarianceMatrixFile = NULL,
  chromosome = NULL,
  generatePermutations = T,
  snp_gene_map_file_name = NULL
)
}
\arguments{
\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{covariates_file_name}{(optional) A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.}

\item{eQTLInputFile}{The empiric pvalues, as generated by runEQTL.}

\item{permutationFileName}{The name of the permutation file to generate (This describes how genotype columns are permuted for each permutation.)}

\item{eQTLOutputFullPermutationFile}{The results after running second round full permutation (ALL snps per gene).}

\item{snpReportFile}{Output a report of SNPs that surpase the empiric pvalues during permutation.}

\item{maxNumIter}{how many iterations should permutation be run for.}

\item{cisDist}{defines the window around each gene to find associated SNPs.}

\item{minNumPassingPermuations}{how many times must the permuted value > empiric value before adaptive permutation stops testing a gene.}

\item{useModel}{The model MatrixEQTL uses.}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{errorCovarianceMatrixFile}{If supplied, a matrix of the relationships between donors, which is calculated as (kinship *2).
This can be a superset of the donors in the other files.}

\item{chromosome}{If provided limit analysis to a single chromosome.}

\item{generatePermutations}{If true, a new permutation file is generated.  If false, the existing one is used.  If false and the file does not exist, throws an error.}

\item{snp_gene_map_file_name}{optional file with columns 'snp' and 'gene' that represents a many-to-many
map.  If present, only the list of SNPs mapped to a gene are eQTL candidates.}
}
\value{
The permuted results data frame.
}
\description{
In the full permutation test, ALL snps for a gene are have their genotypes permuted, and for each SNP it's empiric pvalue is compared
to ALL SNPs for the gene.
}
