% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_PermutationTesting.R
\name{runAdaptiveMultiStepPermutationsFromBaseFileNames}
\alias{runAdaptiveMultiStepPermutationsFromBaseFileNames}
\title{Multi-step permutation analysis}
\usage{
runAdaptiveMultiStepPermutationsFromBaseFileNames(
  base.dir,
  base.name,
  output.dir = base.dir,
  maxNumIter = 1000,
  cisDist = 1e+05,
  minNumPassingPermuations = 5,
  useModel = modelLINEAR,
  minorAlleleFreqThreshold = NULL,
  useQuantileNormalization = F,
  errorCovarianceMatrixFile = NULL,
  chromosome = NULL,
  generatePermutations = T
)
}
\arguments{
\item{base.dir}{The directory data resides in.}

\item{base.name}{The prefix of all files for this data set.}

\item{output.dir}{The directory where output files should be emitted}

\item{maxNumIter}{how many iterations should permutation be run for.}

\item{cisDist}{defines the window around each gene to find associated SNPs.}

\item{minNumPassingPermuations}{how many times must the permuted value > empiric value before adaptive permutation stops testing a gene.}

\item{useModel}{The model MatrixEQTL uses.}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{errorCovarianceMatrixFile}{If supplied, a matrix of the relationships between donors, which is calculated as (kinship *2).
This can be a superset of the donors in the other files.}

\item{chromosome}{If provided limit analysis to a single chromosome.}

\item{generatePermutations}{should a new set of permutations be used, or should existing permutation files be used.}
}
\value{
The results of adaptive permutation (ALL SNPs per gene.)
}
\description{
Multi-step permutation analysis
}
