% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_PermutationTesting.R
\name{runAdaptivePermutationEQTL}
\alias{runAdaptivePermutationEQTL}
\title{Run adaptive permutation testing of eQTL results.}
\usage{
runAdaptivePermutationEQTL(
  SNP_file_name,
  snps_location_file_name,
  expression_file_name,
  gene_location_file_name,
  covariates_file_name = NULL,
  eQTLInputFile,
  eQTLOutputFile,
  snpReportFile = NULL,
  useQuantileNormalization = F,
  minorAlleleFreqThreshold = NULL,
  useModel = modelLINEAR,
  errorCovarianceMatrixFile = NULL,
  maxNumIter = 1e+05,
  cisDist = 1e+05,
  minNumPassingPermuations = 5,
  permutationFileName = NULL,
  geneList = NULL,
  chromosome = NULL,
  snp_gene_map_file_name = NULL
)
}
\arguments{
\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{covariates_file_name}{(optional) A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.}

\item{eQTLInputFile}{The non-permuted eQTL results file, generated by @see runEQTL}

\item{eQTLOutputFile}{The permuted output results file.}

\item{snpReportFile}{A file containing the best SNP for each gene where the permutation result for the gene has a better p-value than the empiric result.}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{useModel}{The model MatrixEQTL uses.}

\item{errorCovarianceMatrixFile}{If supplied, a matrix of the relationships between donors, which is calculated as (kinship *2).
This can be a superset of the donors in the other files.}

\item{maxNumIter}{how many iterations should permutation be run for.}

\item{cisDist}{defines the window around each gene to find associated SNPs.}

\item{minNumPassingPermuations}{how many times must the permuted value > empiric value before adaptive permutation stops testing a gene.}

\item{permutationFileName}{The name of the permutation file to generate (This describes how genotype columns are permuted for each permutation.)}

\item{geneList}{If provided, limit analysis to a subset of genes.}

\item{chromosome}{If provided limit analysis to a single chromosome.}

\item{snp_gene_map_file_name}{optional file with columns 'snp' and 'gene' that represents a many-to-many
map.  If present, only the list of SNPs mapped to a gene are eQTL candidates.}
}
\value{
The permuted results data frame.
}
\description{
This is a highly paramterized version of adaptive permutation that can run either best SNP/gene or gold-standard permutation testing.
}
