% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{runECAVIAR}
\alias{runECAVIAR}
\title{Run CAVIAR/eCAVIAR on a single region as defined by the outputBaseFileName file set.
Generates outputs to the outputBaseFileName with appropriate suffixes.}
\usage{
runECAVIAR(
  CAVIARPath,
  eCAVIARPath,
  outputBaseFileName,
  dryRun = F,
  verbose = F
)
}
\arguments{
\item{CAVIARPath}{The full path to the CAVIAR executable}

\item{eCAVIARPath}{The full path to the eCAVIAR executable}

\item{outputBaseFileName}{The prefix for all files for this region, including the full directory path.}

\item{dryRun}{Print out the commands that would be run for 3rd party tools, but don't execute anything.}

\item{verbose}{More logging during run.}
}
\description{
Run CAVIAR/eCAVIAR on a single region as defined by the outputBaseFileName file set.
Generates outputs to the outputBaseFileName with appropriate suffixes.
}
