% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_Example.R
\name{runEQTL}
\alias{runEQTL}
\title{eQTL discovery.}
\usage{
runEQTL(
  params_file_name,
  cis_eqtl_file_name,
  SNP_file_name,
  snps_location_file_name,
  expression_file_name,
  gene_location_file_name,
  covariates_file_name = NULL,
  useQuantileNormalization = F,
  minorAlleleFreqThreshold = NULL,
  pvOutputThreshold_cis = 1,
  useModel = modelLINEAR,
  cisDist = 1e+05,
  verbose = TRUE,
  noFDRsaveMemory = TRUE,
  snp_gene_map_file_name = NULL,
  errorCovarianceMatrixFile = NULL
)
}
\arguments{
\item{params_file_name}{Matrix_eQTL_main param elements (except multi-valued ones) written here in tabular form.}

\item{cis_eqtl_file_name}{Results of Matrix_eQTL_main written here.}

\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{covariates_file_name}{(optional) A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{pvOutputThreshold_cis}{cis eQTLs must have a pvalue less than or equal to this threshold to be reported.}

\item{useModel}{The model MatrixEQTL uses.}

\item{cisDist}{defines the window around each gene to find associated SNPs.}

\item{verbose}{output detailed information about the eQTL process as it runs (passed to MatrixEQTL::Matrix_eQTL_main)}

\item{noFDRsaveMemory}{passed through to MatrixEQTL::Matrix_eQTL_main.}

\item{snp_gene_map_file_name}{optional file with columns 'snp' and 'gene' that represents a many-to-many
map.  If present, only the list of SNPs mapped to a gene are eQTL candidates.}

\item{errorCovarianceMatrixFile}{If supplied, a matrix of the relationships between donors, which is calculated as (kinship *2).
This can be a superset of the donors in the other files.}
}
\value{
the eQTL result data frame.
}
\description{
Runs Matrix eQTL package on a data set.
}
\details{
This validates files exist, loads up data, runs the analysis, and writes out the results.
Optionally you can filter variants on MAF, or quantile normalize the expression data.
}
\seealso{
\code{\link{Matrix_eQTL_main}}.

MatrixEQTL::Matrix_eQTL_main
}
