% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_Example.R
\name{runEQTLFromBaseFileNames}
\alias{runEQTLFromBaseFileNames}
\title{eQTL discovery with assumed input file names.}
\usage{
runEQTLFromBaseFileNames(
  base.dir,
  base.name,
  useQuantileNormalization = F,
  minorAlleleFreqThreshold = NULL,
  pvOutputThreshold_cis = 1,
  useModel = modelLINEAR,
  cisDist = 1e+05
)
}
\arguments{
\item{base.dir}{The directory data resides in.}

\item{base.name}{The prefix of all files for this data set.}

\item{useQuantileNormalization}{Should the data be quantile normalized?}

\item{minorAlleleFreqThreshold}{Should the SNPs be filtered on minor allele frequency?}

\item{pvOutputThreshold_cis}{cis eQTLs must have a pvalue less than or equal to this threshold to be reported.}

\item{useModel}{The model MatrixEQTL uses.}

\item{cisDist}{defines the window around each gene to find associated SNPs.}
}
\value{
the eQTL result data frame.
}
\description{
run eQTL analysis, using a convenience method to define the file name inputs/outputs to a standard format.
}
\seealso{
\code{\link{runEQTL}}.
}
