% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLUtils.R
\name{runOneCovarMatrixeQTL}
\alias{runOneCovarMatrixeQTL}
\title{Run matrix eQTL on an expression data as the dependent variable to measure the effects of an independent variable encoded by a covariate.}
\usage{
runOneCovarMatrixeQTL(covarName, covars, expData, numPermutations = 1000)
}
\arguments{
\item{covarName}{The covariate name to analyze}

\item{covars}{A data frame of 1 or more covariates.  Each row is a covariate, each column a sample.  The first column "id" contains the name of the covariate.}

\item{expData}{A data frame of expression data, 1 row per gene, 1 column per sample. The first column "id" contains the name of the gene.}

\item{numPermutations}{Runs permutations of the linear regression, permuting the covariate donor labels.}
}
\value{
The empiric results if numPermutations=0, else the permuted results.
}
\description{
Run matrix eQTL on an expression data as the dependent variable to measure the effects of an independent variable encoded by a covariate.
}
