% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegionAndExpressionByGenotype.R
\name{runOneGwasHit}
\alias{runOneGwasHit}
\title{Scan/plot eQTL results by index SNP}
\usage{
runOneGwasHit(
  gwasSNP,
  gene_location_file_name,
  snps_location_file_name,
  expression_file_name,
  SNP_file_name,
  index_snps_file,
  empiric_pvalues,
  outDir
)
}
\arguments{
\item{gwasSNP}{The formatted location name of the GWAS SNP.  This is the chromosome:position:ref:alt, for example
"chr15:78576056:C:T"}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{index_snps_file}{The eQTL permuted results.  This is typically the eigenMT index file generated by the eQTL discovery pipeline - ie: eQTL_eigenMT_results.txt.
This should use the search window size and MAF appropriate for your sample size.}

\item{empiric_pvalues}{The eQTL empiric pvalues file.  This is typically the empric pvalues file from the eQTL discovery pipeline - ie: .eQTL_results.txt.  This
typically uses a more permissive search window size and MAF than the index_snps_file input.}

\item{outDir}{The output directory to generate the PDF report.  The report file is named for both the query SNP name as well as the basename of the input data.}
}
\description{
For a given SNP, search for all genes that were tested by that SNP in the empiric pvalues input file.  For each gene, plot the region
empiric pvalues as a manhatten plot. Also plot the eQTL index SNP and GWAS SNP genotype by expression interaction plot.
}
