% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateGeneCorrelation.R
\name{runOneManyCovars}
\alias{runOneManyCovars}
\title{Run eQTL analysis one iteration to look at many covariates}
\usage{
runOneManyCovars(expDataT, covarsList)
}
\arguments{
\item{expDataT}{The expData matrix, transposed so each row is a donor,
and the gene ID's made into column names.}

\item{covarsList}{The covariates from the matrix transformed into a list of vectors, with
each list element the name of the covariate.}
}
\value{
A data frame containing each gene, and the pvalue for the each covariate as well as the full model.
}
\description{
This uses data that's been pre-processed to save time.
}
