% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateGeneCorrelation.R
\name{runPermutationManyCovars}
\alias{runPermutationManyCovars}
\title{Run permutations on regresssions using multiple covariates}
\usage{
runPermutationManyCovars(
  result,
  expDataT,
  covarsList,
  numPermutations = 1000,
  random.seed = 1,
  numThreads = 1
)
}
\arguments{
\item{result}{The empiric result}

\item{expDataT}{The expData matrix, transposed so each row is a donor,
and the gene ID's made into column names.}

\item{covarsList}{The covariates from the matrix transformed into a list of vectors, with
each list element the name of the covariate.}

\item{numPermutations}{The number of permutations to run}

\item{random.seed}{The random seed}

\item{numThreads}{Number of threads used for processing.}
}
\value{
A data frame of permutation results
}
\description{
Run permutations on regresssions using multiple covariates
}
