% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{For a data frame of SNPs in both eQTL and GWAS data, run the caviar fine mapping pipeline.}
\usage{
runPipeline(
  inSNPFile,
  manifestFile,
  vcfFile,
  gene_location_file_name,
  outDir,
  bcfToolsLocation,
  ldScoreBcftoolsPluginLocation,
  CAVIARPath,
  eCAVIARPath,
  dryRun = F,
  verbose = T,
  zScoreAbsValues = F
)
}
\arguments{
\item{inSNPFile}{A file containing a combined SNP data frame of eQTL + GWAS data, generated by buildGWASeQTLSNPSets}

\item{manifestFile}{A file containing each inSNPFile number, and the gene it corresponds to.}

\item{vcfFile}{A VCF file to extract LD data from for this region.  Suggest using a population from thousand genomes or similar.}

\item{gene_location_file_name}{A file name containing the coordinates of genes.  Has 4 columns: geneid, chr, s1, s2.  S1 is the start, S2 the end of the gene.}

\item{outDir}{Where should data be output to?}

\item{bcfToolsLocation}{The full path to the bcftools executable.}

\item{ldScoreBcftoolsPluginLocation}{The ful path to the ldScore plugin for bcftools.}

\item{CAVIARPath}{The full path to the CAVIAR executable}

\item{eCAVIARPath}{The full path to the eCAVIAR executable}

\item{dryRun}{Print out the commands that would be run for 3rd party tools, but don't execute anything.}

\item{verbose}{More logging during run.}

\item{zScoreAbsValues}{Use the unsigned Z scores for inputs to CAVIAR/eCAVIAR}
}
\description{
For a data frame of SNPs in both eQTL and GWAS data, run the caviar fine mapping pipeline.
}
\seealso{
buildGWASeQTLSNPSets
}
