% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{runPipelineMany}
\alias{runPipelineMany}
\title{Title}
\usage{
runPipelineMany(
  inSNPFilePath,
  manifestFile,
  vcfFile,
  outDir,
  bcfToolsLocation,
  ldScoreBcftoolsPluginLocation,
  CAVIARPath,
  eCAVIARPath,
  dryRun = F,
  verbose = F
)
}
\arguments{
\item{inSNPFilePath}{A directory containing one or more combined SNP data frames of eQTL + GWAS data, generated by buildGWASeQTLSNPSets}

\item{manifestFile}{A file containing each inSNPFile number, and the gene it corresponds to.}

\item{vcfFile}{A VCF file to extract LD data from for this region.  Suggest using a population from thousand genomes or similar.}

\item{outDir}{Where should data be output to?}

\item{bcfToolsLocation}{The full path to the bcftools executable.}

\item{ldScoreBcftoolsPluginLocation}{The ful path to the ldScore plugin for bcftools.}

\item{CAVIARPath}{The full path to the CAVIAR executable}

\item{eCAVIARPath}{The full path to the eCAVIAR executable}

\item{dryRun}{Print out the commands that would be run for 3rd party tools, but don't execute anything.}

\item{verbose}{More logging during run.}
}
\description{
Title
}
