% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{runSinglePhenotypePermutationsAdaptive}
\alias{runSinglePhenotypePermutationsAdaptive}
\title{Run permutation testing on the empiric results.}
\usage{
runSinglePhenotypePermutationsAdaptive(
  snps,
  snpspos,
  p,
  empiricResults,
  numPermutations = 10000,
  minNumPassingPermuations = 15,
  output_file_name_cis
)
}
\arguments{
\item{snps}{Matrix eQTL package SlicedData object containing the genotype matrix (see: SNP_file_name)}

\item{snpspos}{data table containingthe SNP locations (see: snps_location_file_name)}

\item{p}{Matrix eQTL package SlicedData object containing the phenotype to test}

\item{empiricResults}{Dataframe containing the matrix eQTL empiric results for all SNPs in the snps object.}

\item{numPermutations}{The maximum number of permutations run on the result set.  Permutations are adaptive - for SNPs
where the permuted results quickly exceed the empiric pvalues, a smaller number of permutations are run to conserve time.}

\item{minNumPassingPermuations}{How many times do we need to see random data exceed the empiric pvalue to be satisfied with the permuted p?}

\item{output_file_name_cis}{If not null, write intermediate outputs to this file.}
}
\description{
Use adaptive permutation testing - test each SNP 100 times, then eliminate all results where the permuted result was
significant at least minNumPassingPermuations times.
}
