% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{runSinglePhenotypePermutationsSimple}
\alias{runSinglePhenotypePermutationsSimple}
\title{Run permutation testing on the empiric results for the subset of SNPs with marginally significant p-values}
\usage{
runSinglePhenotypePermutationsSimple(
  snps,
  snpspos,
  p,
  empiricResults,
  permutationPvalueThreshold = 0.001,
  numPermutations = 10000
)
}
\arguments{
\item{numPermutations}{The maximum number of permutations run on the result set.  Permutations are adaptive - for SNPs
where the permuted results quickly exceed the empiric pvalues, a smaller number of permutations are run to conserve time.}
}
\description{
This is a pretty brute force approach without adaptive permutation testing.
}
