% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLPlots.R
\name{scatterPlotPvaluesForTwoMethods}
\alias{scatterPlotPvaluesForTwoMethods}
\title{Scatter plot of p-values for (SNP, gene) that overlap between 2 sets of eQTL_permuted_results}
\usage{
scatterPlotPvaluesForTwoMethods(
  method1Path,
  method2Path,
  method1Label = basename(method1Path),
  method2Label = basename(method2Path),
  title = NA,
  outFile = NULL
)
}
\arguments{
\item{method1Path}{tab-separated file with columns (SNP, gene, p.value) (and maybe other columns).  Plotted on x-axis.}

\item{method2Path}{tab-separated file with columns (SNP, gene, p.value) (and maybe other columns).  Plotted on y-axis.}

\item{method1Label}{x-axis label.}

\item{method2Label}{y-axis label.}

\item{title}{plot title}

\item{outFile}{If specified, create PDF with plot}
}
\description{
Scatter plot of p-values for (SNP, gene) that overlap between 2 sets of eQTL_permuted_results
}
