% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SexBiasedGeneExpression.R
\name{sexBiasedGeneExpression}
\alias{sexBiasedGeneExpression}
\title{Find genes with biased expression based SEX or FRACTION_X}
\usage{
sexBiasedGeneExpression(
  expression_file_name,
  covariates_file_name,
  gene_location_file_name = NULL,
  contigGroupsFile = NULL,
  outFile = NULL,
  numPermutations = 10000,
  renormAutosomeExpression = T,
  random.seed = 1234
)
}
\arguments{
\item{expression_file_name}{A matrix of expression data, 1 row per gene, 1 column per sample.}

\item{covariates_file_name}{A matrix of 1 or more covariates.  Each row is a covariate, each column a sample.  Contains a row with the donors' gender.}

\item{gene_location_file_name}{The location of each gene.  Same number of lines as the expression_file_name, encodes the gene name, chromosome, start, end.}

\item{contigGroupsFile}{A YAML file containing each contig name and groups it belongs to.  In this case the contigs should belong to the groups autosome and chrX.  If provided, renormalizes autosome expression data to not include the x chromosomne.}

\item{outFile}{The report of each gene tested and the p-values and effect sizes.}

\item{numPermutations}{How many permutations should be run to generate permuted p-values}

\item{renormAutosomeExpression}{If a contigGroupsFile and this is true, then renormalize expression of the autosome to the sum of the autosome genes (instead of all genes)}

\item{random.seed}{A random seed to ensure reproducibility across runs of the same data.}
}
\description{
Converts FRACTION_X into X reactivation.  This normalizes fraction_X by the mean FRACTION_X of individuals with SEX=1.
This makes all SEX=1 individuals close to 1.
}
\details{
Analysis is broken down into a few steps:
(Optionally) Normalize autosome expression to expression/sum(expression) [this removes the X chromosome influence on normalization]
1) For the subset of XX individuals, regression : expression ~ X reactivation
1a) Run permutation testing by permuting expression labels
2) Normalize XX individuals expression to account for X reactivation covariate.
3) Test the effects of SEX on normalized expression (expression [conditional on X reactivation] ~ SEX)
3a) Run permutations on SEX regression
}
