% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{singleCellEqtl}
\alias{singleCellEqtl}
\title{Discover eQTL genotype interactions with latent factors}
\usage{
singleCellEqtl(
  snpFile,
  h5SeuratFile,
  cellMetaDataFile,
  donorCovariatesFile,
  permutedPseudobulkEqtlResultsFile,
  usageFile = NULL,
  qvalueThreshold = 0.05,
  usePeerFactors = T,
  eQTLGeneList = NULL,
  chromosome = NULL,
  factorName = c("FACTOR_2"),
  nPerm = 0,
  outDetailsFile = NULL,
  outSummaryFile = NULL
)
}
\arguments{
\item{snpFile}{A SNP genotype matrix as generated by the eQTL pipeline}

\item{h5SeuratFile}{A file containing single cell expression data in h5 Seruat format}

\item{cellMetaDataFile}{A file containing a matrix of cell level metadata}

\item{donorCovariatesFile}{A file containing donor known covariates}

\item{permutedPseudobulkEqtlResultsFile}{The result file from the eQTL pipeline containg the permuted eQTL results generated by a pseudobulk analysis.}

\item{usageFile}{A file containing a matrix of latent factor scores}

\item{qvalueThreshold}{Filter snp/gene interactions in the permutedPseudobulkEqtlResultsFile}

\item{usePeerFactors}{If true, include the donor covariates from that match "PEER_"}

\item{eQTLGeneList}{An explicit list of genes to further filter the set of SNP/Gene interactions being tested}

\item{chromosome}{A single chromosome to filter the SNP/Gene interactions on}

\item{factorName}{Selects the factor(s) from the usageFile to analyze.  If null, regress all latent factors}

\item{nPerm}{The number of permutations to run to generate a permuted pvalue.  This has two functions:
1) It generates an permuted pvalue where the (number of permuted results > empiric result)+1 / total results +1 = permuted pvalue.
2) The distributon of permuted pvalues is used to fit a beta distribution to compute a final permuted pvalue, in the
same manner as fastQTL.}

\item{outDetailsFile}{For each SNP/Gene pair, ouputs all the coefficents in the fitted main model,
as well as the pvalue from the comparison of the fitted to the null model. One coefficient per line.}

\item{outSummaryFile}{For each SNP/Gene pair, outputs the genotype, factor, and genotype:factor interaction
as well as the pvalue from the comparison of the fitted to the null model.  One result per line.}

\item{quantizationErrorThreshold}{If the cell level metadata contains the column QUANT_ERROR, the data has been previously compressed, and will be filtered to cells with values less than this score.}
}
\description{
For a given set of eQTLs, fit those SNP/gene pairs using a poisson regression with random effects for the donor
and batch.  The set of SNP/gene pairs tested is defined by the permutedPseudobulkEqtlResultsFile, the qvalue threshold to filter
that file, the chromosome to optionally further filter that gene list to a single chromosome, and optionally an eQTLGeneList
that contains an explicit set of genes to test.
}
\details{
If usageFile is null, run the standard eQTL analysis without the latent factor interaction.
If the usage file is provided, runs the genotype:latent factor interaction analysis
}
