% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellEqtl.R
\name{singleCellEqtlPermuted}
\alias{singleCellEqtlPermuted}
\title{Permutes data to calculate an FDR of the non-permuted results}
\usage{
singleCellEqtlPermuted(
  snpFile,
  h5SeuratFile,
  cellMetaDataFile,
  donorCovariatesFile,
  permutedPseudobulkEqtlResultsFile,
  usageFile = NULL,
  qvalueThreshold = 0.05,
  usePeerFactors = T,
  eQTLGeneList,
  chromosome = NULL,
  factorName = c("FACTOR_2"),
  numClustersPerDonor = NULL,
  outPermutedDetailsFile = NULL,
  outPermutedSummaryFile = NULL
)
}
\arguments{
\item{snpFile}{A SNP genotype matrix as generated by the eQTL pipeline}

\item{h5SeuratFile}{A file containing single cell expression data in h5 Seruat format}

\item{cellMetaDataFile}{A file containing a matrix of cell level metadata}

\item{donorCovariatesFile}{A file containing donor known covariates}

\item{permutedPseudobulkEqtlResultsFile}{The result file from the eQTL pipeline containg the permuted eQTL results generated by a pseudobulk analysis.}

\item{usageFile}{A file containing a matrix of latent factor scores}

\item{qvalueThreshold}{Filter snp/gene interactions in the permutedPseudobulkEqtlResultsFile}

\item{usePeerFactors}{If true, include the donor covariates from that match "PEER_"}

\item{eQTLGeneList}{An explicit list of genes to further filter the set of SNP/Gene interactions being tested}

\item{chromosome}{A single chromosome to filter the SNP/Gene interactions on}

\item{factorName}{Selects the factor(s) from the usageFile to analyze.  If null, regress all latent factors}

\item{outPermutedDetailsFile}{For each SNP/Gene pair, ouputs all the coefficents in the fitted main model,
as well as the pvalue from the comparison of the fitted to the null model. One coefficient per line.}

\item{outPermutedSummaryFile}{For each SNP/Gene pair, outputs the genotype, factor, and genotype:factor interaction
as well as the pvalue from the comparison of the fitted to the null model.  One result per line.}
}
\description{
If no usageFile is provided, test the effect of genotype on expression.  The permutation
test permutes the donor genotype, so all cells that have a particular donor genotype will
retain the same genotype, but that genotype may be different.
}
\details{
If a usage file is provided, test the effect of the genotype:factor interaction.  The
permutation test permutes the factor score for each cell.
}
