% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{singlePhenotypePermutation}
\alias{singlePhenotypePermutation}
\title{Permute the genotype data, then retest the genotype-phenotype relationship}
\usage{
singlePhenotypePermutation(snps, p, empiricResultsTested, verbose = T)
}
\arguments{
\item{snps}{Matrix eQTL package SlicedData object containing the genotype matrix (see: SNP_file_name)}

\item{p}{Matrix eQTL package SlicedData object containing the phenotype to test}

\item{empiricResultsTested}{The set of eQTL results to permute.}
}
\description{
For each SNP, increment the number of tests by 1.
For each SNP where the permuted test P < empiric test P, increment excceds_empiric_p by 1.
}
