% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{singlePhenotypePermutationFaster}
\alias{singlePhenotypePermutationFaster}
\title{Permute the phenotype data, then retest the genotype-phenotype relationship}
\usage{
singlePhenotypePermutationFaster(snps, p, empiricResultsTested, verbose = T)
}
\description{
It's possible the single permutation is being slowed down by the pass by value semantics
of R when the genotype matrix is sampled and thus changed.  Since there's only
a single phenotype, permute that instead.
}
\details{
For each SNP, increment the number of tests by 1.
For each SNP where the permuted test P < empiric test P, increment excceds_empiric_p by 1.
}
