% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePhenotypePlots.R
\name{singlePhenotypeStandardAnalysis}
\alias{singlePhenotypeStandardAnalysis}
\title{Standard Analysis for a single phenotype GWAS scan}
\usage{
singlePhenotypeStandardAnalysis(
  gwasResult,
  variantLocationFile,
  population = "EUR",
  pValueThreshold = 1e-05
)
}
\arguments{
\item{gwasResult}{A file containing the permuted output of the testSinglePhenotype function}

\item{variantLocationFile}{The location of each SNP.  Encodes the snp name, chromosome, position, rsID.}

\item{population}{If supplied, fetches LD for the SNP from this thousand genomes population. See the LD link website: https://ldlink.nci.nih.gov/}

\item{pValueThreshold}{Pvalue threshold to define which loci are plotted.}
}
\description{
Generates manhatten plot for empiric and permuted p-values
Generates a region plot for each locus that contains a SNP above the pvalue threshold
Generates a genotype vs phenotype plot for the best SNP in the locus
}
\seealso{
DropSeq.eqtl::testSinglePhenotype
}
