% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePermutationResults.R
\name{sortByChromosomePosSequenceDictionary}
\alias{sortByChromosomePosSequenceDictionary}
\title{Sort the permuted output data frame by SNP chromosome and position ordering}
\usage{
sortByChromosomePosSequenceDictionary(result, sdFile)
}
\arguments{
\item{result}{The permuted results per chromosome bound into a single data frame.  This must contain a SNP
column that is formatted as chromosome:position:ref_allele:alt_allele.}

\item{sdFile}{A file detailing the expected order of chromosomes to perserve genomic ordering for
the reference, generated by Picard's CreateSequenceDictionary.}
}
\value{
The input data frame, sorted in genomic order.
}
\description{
Uses a sequence dictionary file (see Picard's CreateSequenceDictionary) to determine the order of
chromosomes in the output.
}
