% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLUtils.R
\name{splitResultsByContigGroup}
\alias{splitResultsByContigGroup}
\title{Split a data set into a list of data frames by defined contig groups}
\usage{
splitResultsByContigGroup(df, contigColumn = "chr", contigGroupsFile)
}
\arguments{
\item{df}{A data frame that has a column defining the contig name}

\item{contigColumn}{The name of the column that has contig names}

\item{contigGroupsFile}{A YAML file containing the groups contigs belong to}
}
\value{
A list of data frames containing the input information, split into new groups.
}
\description{
Split a data set into a list of data frames by defined contig groups
}
