% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{summarizeSignTestManyExperiments}
\alias{summarizeSignTestManyExperiments}
\title{Summarize results across many eQTL experiments.}
\usage{
summarizeSignTestManyExperiments(
  experimentNames,
  permutedQTLFiles,
  unfilteredQTLFiles,
  geneDescriptionsFile = NULL,
  qvalueThreshold = 0.05,
  outGeneLevelSummaryFile = NULL,
  outExperimentLevelSummaryFile = NULL,
  filterChromosomeList = NULL
)
}
\arguments{
\item{experimentNames}{The names of the experiment to appear in the columns of the outputs.}

\item{permutedQTLFiles}{A vector of eQTL result file that has gone through permutation (or other gene-level correction) and has a q-value.}

\item{unfilteredQTLFiles}{A vector of eQTL result file run on all SNPs on the data set.  This data set should contain all SNP/Gene pairs in the permutedQTLFile,
even if they wouldn't normally be included in the data set.}

\item{geneDescriptionsFile}{A file containing gene names in column 2 and gene descriptions in column 3.}

\item{qvalueThreshold}{A minimum q-value threshold for a gene from the permutedQTLFile to be included in the result set.}

\item{outGeneLevelSummaryFile}{Output file containing the gene-level results}

\item{outExperimentLevelSummaryFile}{Output file containing the experiment level sign test results.}

\item{filterChromosomeList}{remove chromosomes in this vector.  Useful when you want to filter out the X chromosome.}
}
\description{
Captures the set of genes with a qvalue below the threshold in any experiment, and the
q-values for each experiment.  Runs the sign test on the appropriate SNP for each experiment.
}
