% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{testEffectsSignsMatch}
\alias{testEffectsSignsMatch}
\title{Performs a sign test on two data sets to see if SNP/Gene combinations show the same direction of effect.}
\usage{
testEffectsSignsMatch(
  unfilteredQTLFile,
  permutedQTLFile,
  qvalueThreshold = 0.05,
  verbose = T,
  filterChromosomeList = NULL,
  chunk_size = 1e+06,
  outMetricsFile = NULL,
  flipAlleles = T,
  stripENSGSuffix = T
)
}
\arguments{
\item{unfilteredQTLFile}{An eQTL result file run on all SNPs on the data set.  This data set should contain all SNP/Gene pairs in the permutedQTLFile,
even if they wouldn't normally be included in the data set.}

\item{permutedQTLFile}{An eQTL result file that has gone through permutation (or other gene-level correction) and has a q-value.}

\item{qvalueThreshold}{A minimum q-value threshold for a gene from the permutedQTLFile to be included in the result set.}

\item{verbose}{Should "helpful" information be printed as function is run?}

\item{filterChromosomeList}{remove chromosomes in this vector.  Useful when you want to filter out the X chromosome.}

\item{chunk_size}{how much of the unfiltered eQTL file should be loaded at a time.  These files can be quite large.}

\item{outMetricsFile}{If not null, writes out a file with statistics on the run (number/fraction eGenes passing sign test)}

\item{flipAlleles}{If true, looks for instances where the reference and alternate alleles in the two data sets are reversed.  Sets
the reference and alternate allele in the unfilteredQTLFile to be the same as the permutedQTLFile, and where the alleles are flipped
the beta is multiplied by -1.}

\item{stripENSGSuffix}{Should the suffix be removed from ENSG IDs if they are used as the gene names?  For example
an ENSG can have a version number: ENSG00000000460.12 where 12 is the version.  If this is set to be true the ".12" is removed.}
}
\value{
A list containing 2 elements: 1) A data frame containing a list of genes/SNPs that passed the FDR threshold, their q-value in permutedQTLFile,
the direction of the effect in the permutedQTLFile, and the direction in the unfilteredQTLFile.
2) A calculation of the fraction of snp/gene pairs where the sign of the two results agree.
}
\description{
Performs a sign test on two data sets to see if SNP/Gene combinations show the same direction of effect.
}
