% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignTest.R
\name{testEffectsSignsMatchMany}
\alias{testEffectsSignsMatchMany}
\title{Run the sign test on multiple data sets and aggregate the results.}
\usage{
testEffectsSignsMatchMany(
  permutedQTLFile,
  unfilteredQTLFiles,
  qvalueThreshold = 0.05,
  verbose = T,
  filterChromosomeList = NULL
)
}
\arguments{
\item{permutedQTLFile}{An eQTL result file that has gone through permutation (or other gene-level correction) and has a q-value.}

\item{unfilteredQTLFiles}{A list of unfiltered eQTL files}

\item{qvalueThreshold}{A minimum q-value threshold for a gene from the permutedQTLFile to be included in the result set.}

\item{verbose}{Should "helpful" information be printed as function is run?}

\item{filterChromosomeList}{remove chromosomes in this vector.  Useful when you want to filter out the X chromosome.}
}
\value{
A data frame containing the gene and a column containing the sign test data across experiments.  This is
the snpName : sign of experiment.  For example SNP123:+/+/+/+ would indicate 4 experiments where the direction is the same for all 4.
Also returns a single row data frame with the fraction of the genes from the permutedQTLFile where the sign matches in the unfiltered experiment.
}
\description{
Given one set of genes with permuted p-values, scan other data sets to see if their sign agrees.  Aggregate results of multiple unfiltered data sets into a single result.
}
