% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{testSinglePhenotype}
\alias{testSinglePhenotype}
\title{Test a single phenotype against all genotypes to find associations}
\usage{
testSinglePhenotype(
  SNP_file_name,
  snps_location_file_name,
  phenotypeFileName,
  CNV_file_name = NULL,
  CNV_location_file_name = NULL,
  output_file_name_cis,
  numPermutations = 10000,
  minNumPassingPermuations = 15,
  chromosome = NULL
)
}
\arguments{
\item{SNP_file_name}{A matrix of genotypes, 1 row per SNP, 1 column per sample.  Encoded as 0,1,2 copies of the alternate allele.}

\item{snps_location_file_name}{The location of each SNP.  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, position.}

\item{phenotypeFileName}{A matrix of phenotypes, 1 row per phenotype, 1 column per sample.}

\item{CNV_file_name}{A matrix of CNV genotypes, 1 row per SNP, 1 column per sample.  Encoded as whole or dosage copy number.  Optional.}

\item{CNV_location_file_name}{The location of each CNV  Same number of lines as the SNP_file_name, encodes the snp name, chromosome, start position, end position.}

\item{output_file_name_cis}{The eQTL result file.}

\item{numPermutations}{The maximum number of permutations run on the result set.  Permutations are adaptive - for SNPs
where the permuted results quickly exceed the empiric pvalues, a smaller number of permutations are run to conserve time.}

\item{minNumPassingPermuations}{How many times do we need to see random data exceed the empiric pvalue to be satisfied with the permuted p?}

\item{chromosome}{If provided limit analysis to a single chromosome.}
}
\description{
Take in a matrix of genotypes, snp locations, and a single phenotype formatted like an expression matrix input.
Optionally, supply a matrix of CNV genotypes and CNV locations.
Runs all genotypes/CNVs against the single expression vector.
}
