% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{unionGwasEQTL}
\alias{unionGwasEQTL}
\title{Union the GWAS and eQTL effects.
Find SNPs common bot both data sets genome wide.   Check the reference and alternate
alleles for each data set, and make them match in both data sets, flipping the odds ratio
of the GWAS data as appropriate so the eQTL and GWAS use the same allele to determine direction
of effect.
For SNPs present in only 1 data set, retain those "as is".}
\usage{
unionGwasEQTL(gwas, eQTL)
}
\arguments{
\item{gwas}{The GWAS data frame}

\item{eQTL}{The eQTL data frame}
}
\value{
A list containing the unioned/modified GWAS and eQTL data frames.
}
\description{
Union the GWAS and eQTL effects.
Find SNPs common bot both data sets genome wide.   Check the reference and alternate
alleles for each data set, and make them match in both data sets, flipping the odds ratio
of the GWAS data as appropriate so the eQTL and GWAS use the same allele to determine direction
of effect.
For SNPs present in only 1 data set, retain those "as is".
}
