% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_Example.R
\name{validateAndFilterErrorCovarianceMatrix}
\alias{validateAndFilterErrorCovarianceMatrix}
\title{Validate that if the errorCovariance is not null, the donors are in the same order as the expression data.
If the errorCovariance donors are a superset of the donors in the genes data set, restrict the error covariance to be
the same set of donors.
If the errorCovariance does not include all donors in the gene object, but is supplied, this will generate an error
and analysis will not continue.}
\usage{
validateAndFilterErrorCovarianceMatrix(errorCovariance, gene)
}
\arguments{
\item{errorCovariance}{The error covariance, which is either length 0, or a matrix with column names that should match the expression data}

\item{gene}{A SlicedData object containing the expression data}
}
\value{
The error errorCovariance matrix if it is valid.  If invalid, return NULL.
}
\description{
Validate that if the errorCovariance is not null, the donors are in the same order as the expression data.
If the errorCovariance donors are a superset of the donors in the genes data set, restrict the error covariance to be
the same set of donors.
If the errorCovariance does not include all donors in the gene object, but is supplied, this will generate an error
and analysis will not continue.
}
