% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSinglePhenotype.R
\name{validatePhenotypeGenotypeInputs}
\alias{validatePhenotypeGenotypeInputs}
\title{Checks to see that all genotypes have a phenotype}
\usage{
validatePhenotypeGenotypeInputs(snps, p)
}
\arguments{
\item{snps}{A matrix of SNP genotypes (or any other data that has donors in columns)}

\item{p}{A matrix of phenotypes}
}
\value{
the phenotype matrix reordered to be have the same ordering as the genotype matrix.
}
\description{
Reorders phenotype data if necessary to be in the same order as genotype data.
}
