% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_eQTL_CommonFunctions.R
\name{validateSingleFileExists}
\alias{validateSingleFileExists}
\title{Check if a file exists, and if not print a nice error message and stop.}
\usage{
validateSingleFileExists(x)
}
\arguments{
\item{x}{The file to check.  If NULL this is a no-op.}
}
\description{
Check if a file exists, and if not print a nice error message and stop.
}
