% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAVIAR_pipeline.R
\name{writeECAVIARInputs}
\alias{writeECAVIARInputs}
\title{Write out the inputs to CAVIAR/eCAVIAR to perform fine mapping on GWAS/EQTL data sets.
Outputs have no headers.
GWAS/eQTL has 2 columns: snp id and Z score.  Tab separated.
LD block has no headers, square matrix of LD scores.  Space separated.}
\usage{
writeECAVIARInputs(
  dfBoth,
  ldBlockBoth,
  dfGWAS,
  ldBlockGWAS,
  dfEQTL,
  ldBlockEQTL,
  outputBaseFileName,
  zScoreAbsValues = F
)
}
\arguments{
\item{dfBoth}{A data frame containing the GWAS/eQTL SNP data for a region}

\item{ldBlockBoth}{The LD matrix for the intersect of the GWAS / eQTL region, containing the same SNPs as the dfBoth parameter, in the same order.}

\item{dfGWAS}{A data frame containing the GWAS SNP data for a region}

\item{ldBlockGWAS}{The LD matrix for the intersect of the GWAS region, containing the same SNPs as the dfGWAS parameter, in the same order.}

\item{dfEQTL}{A data frame containing the eQTL SNP data for a region.}

\item{ldBlockEQTL}{The LD matrix for the intersect of the eQTL region, containing the same SNPs as the dfEQTL parameter, in the same order.}

\item{outputBaseFileName}{The prefix for all files for this region, including the full directory path.}

\item{zScoreAbsValues}{If true, use the absolute values of the Z-scores in the LD matrix and gwas/eQTL Zscores.}
}
\description{
Write out the inputs to CAVIAR/eCAVIAR to perform fine mapping on GWAS/EQTL data sets.
Outputs have no headers.
GWAS/eQTL has 2 columns: snp id and Z score.  Tab separated.
LD block has no headers, square matrix of LD scores.  Space separated.
}
